% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{add_error_pointwise_1d}
\alias{add_error_pointwise_1d}
\title{Add pointwise Gaussian noise to functional data}
\usage{
add_error_pointwise_1d(data, sd, seed)
}
\description{
Add pointwise Gaussian noise to functional data
}
