% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_align_to_target}
\alias{alignment_align_to_target}
\title{Align all curves to a target curve}
\usage{
alignment_align_to_target(data, target, argvals)
}
\description{
Align all curves to a target curve
}
