% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_amplitude_dist}
\alias{alignment_amplitude_dist}
\title{Amplitude self-distance matrix}
\usage{
alignment_amplitude_dist(data, argvals, lambda)
}
\description{
Amplitude self-distance matrix
}
