% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_bayesian_pair}
\alias{alignment_bayesian_pair}
\title{Bayesian pairwise alignment via pCN MCMC on the Hilbert sphere}
\usage{
alignment_bayesian_pair(
  f1,
  f2,
  argvals,
  n_samples,
  burn_in,
  step_size,
  proposal_variance,
  seed
)
}
\description{
Bayesian pairwise alignment via pCN MCMC on the Hilbert sphere
}
