% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_closed_distance}
\alias{alignment_closed_distance}
\title{Elastic distance between two closed curves (optimizes over rotations)}
\usage{
alignment_closed_distance(f1, f2, argvals, lambda)
}
\description{
Elastic distance between two closed curves (optimizes over rotations)
}
