% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_closed_pair}
\alias{alignment_closed_pair}
\title{Elastic alignment for closed (periodic) curves with rotation search}
\usage{
alignment_closed_pair(f1, f2, argvals, lambda)
}
\description{
Elastic alignment for closed (periodic) curves with rotation search
}
