% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_compose_warps}
\alias{alignment_compose_warps}
\title{Compose two warping functions}
\usage{
alignment_compose_warps(gamma1, gamma2, argvals)
}
\description{
Compose two warping functions
}
