% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_constrained}
\alias{alignment_constrained}
\title{Elastic alignment with landmark constraints}
\usage{
alignment_constrained(
  f1,
  f2,
  argvals,
  landmark_targets,
  landmark_sources,
  lambda
)
}
\description{
Elastic alignment with landmark constraints
}
