% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_curve_geodesic}
\alias{alignment_curve_geodesic}
\title{Geodesic path between two curves in the elastic metric}
\usage{
alignment_curve_geodesic(f1, f2, argvals, n_points, lambda)
}
\description{
Geodesic path between two curves in the elastic metric
}
