% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_decomposition}
\alias{alignment_decomposition}
\title{Elastic phase-amplitude decomposition}
\usage{
alignment_decomposition(f1, f2, argvals, lambda)
}
\description{
Elastic phase-amplitude decomposition
}
