% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_diagnose_pairwise_rust}
\alias{alignment_diagnose_pairwise_rust}
\title{Diagnose a single pairwise alignment}
\usage{
alignment_diagnose_pairwise_rust(f1, f2, gamma, f_aligned, distance, argvals)
}
\description{
Diagnose a single pairwise alignment
}
