% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_diagnose_rust}
\alias{alignment_diagnose_rust}
\title{Diagnose alignment quality for all curves after Karcher mean computation}
\usage{
alignment_diagnose_rust(
  data,
  karcher_mean,
  karcher_mean_srsf,
  gammas,
  aligned_data,
  n_iter,
  converged,
  argvals
)
}
\description{
Diagnose alignment quality for all curves after Karcher mean computation
}
