% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_elastic_depth}
\alias{alignment_elastic_depth}
\title{Elastic depth (amplitude + phase + combined)}
\usage{
alignment_elastic_depth(data, argvals, lambda)
}
\description{
Elastic depth (amplitude + phase + combined)
}
