% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_elastic_distance}
\alias{alignment_elastic_distance}
\title{Elastic (Fisher-Rao) distance between two curves}
\usage{
alignment_elastic_distance(f1, f2, argvals)
}
\description{
Elastic (Fisher-Rao) distance between two curves
}
