% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_elastic_outlier}
\alias{alignment_elastic_outlier}
\title{Elastic outlier detection using Tukey fence on elastic distances}
\usage{
alignment_elastic_outlier(data, argvals, lambda, alpha, use_median)
}
\description{
Elastic outlier detection using Tukey fence on elastic distances
}
