% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_elastic_pair}
\alias{alignment_elastic_pair}
\title{Elastic alignment of one curve to another}
\usage{
alignment_elastic_pair(f1, f2, argvals)
}
\description{
Elastic alignment of one curve to another
}
