% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_elastic_pair_penalized_rust}
\alias{alignment_elastic_pair_penalized_rust}
\title{Penalized elastic alignment with configurable penalty type}
\usage{
alignment_elastic_pair_penalized_rust(f1, f2, argvals, penalty_type, lambda)
}
\description{
Penalized elastic alignment with configurable penalty type
}
