% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_gauss_model}
\alias{alignment_gauss_model}
\title{Gaussian generative model: sample random curves from aligned data}
\usage{
alignment_gauss_model(
  mean_curve,
  mean_srsf,
  gammas,
  aligned_data,
  argvals,
  ncomp,
  n_samples,
  seed
)
}
\description{
Gaussian generative model: sample random curves from aligned data
}
