% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_horiz_fpns}
\alias{alignment_horiz_fpns}
\title{Horizontal Functional Principal Nested Spheres (FPNS) for phase variability}
\usage{
alignment_horiz_fpns(
  mean_curve,
  mean_srsf,
  gammas,
  aligned_data,
  argvals,
  ncomp
)
}
\description{
Horizontal Functional Principal Nested Spheres (FPNS) for phase variability
}
