% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_joint_gauss_model}
\alias{alignment_joint_gauss_model}
\title{Joint Gaussian generative model preserving amplitude-phase correlation}
\usage{
alignment_joint_gauss_model(
  mean_curve,
  mean_srsf,
  gammas,
  aligned_data,
  argvals,
  ncomp,
  n_samples,
  balance_c,
  seed
)
}
\description{
Joint Gaussian generative model preserving amplitude-phase correlation
}
