% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_karcher_mean}
\alias{alignment_karcher_mean}
\title{Karcher (Fréchet) mean in elastic metric}
\usage{
alignment_karcher_mean(data, argvals, max_iter, tol)
}
\description{
Karcher (Fréchet) mean in elastic metric
}
