% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_karcher_mean_closed}
\alias{alignment_karcher_mean_closed}
\title{Karcher mean for closed (periodic) curves}
\usage{
alignment_karcher_mean_closed(data, argvals, max_iter, tol, lambda)
}
\description{
Karcher mean for closed (periodic) curves
}
