% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_karcher_median}
\alias{alignment_karcher_median}
\title{Karcher median (Weiszfeld algorithm in elastic metric)}
\usage{
alignment_karcher_median(data, argvals, max_iter, tol, lambda, trim_fraction)
}
\description{
Karcher median (Weiszfeld algorithm in elastic metric)
}
