% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_lambda_cv_rust}
\alias{alignment_lambda_cv_rust}
\title{Lambda cross-validation for elastic alignment regularisation}
\usage{
alignment_lambda_cv_rust(data, argvals, lambdas, n_folds, max_iter, tol, seed)
}
\description{
Lambda cross-validation for elastic alignment regularisation
}
