% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_multires_pair}
\alias{alignment_multires_pair}
\title{Multi-resolution elastic alignment (coarse DP + fine gradient refinement)}
\usage{
alignment_multires_pair(
  f1,
  f2,
  argvals,
  coarsen_factor,
  n_refine_steps,
  step_size,
  lambda
)
}
\description{
Multi-resolution elastic alignment (coarse DP + fine gradient refinement)
}
