% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_orbit_representative_rust}
\alias{alignment_orbit_representative_rust}
\title{Compute the orbit representative of a curve in a quotient space}
\usage{
alignment_orbit_representative_rust(f, argvals, quotient)
}
\description{
Compute the orbit representative of a curve in a quotient space
}
