% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_pairwise_consistency}
\alias{alignment_pairwise_consistency}
\title{Pairwise alignment consistency}
\usage{
alignment_pairwise_consistency(data, argvals, lambda, max_triplets)
}
\description{
Pairwise alignment consistency
}
