% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_partial_match}
\alias{alignment_partial_match}
\title{Elastic partial matching of a template within a longer target curve}
\usage{
alignment_partial_match(
  template,
  target,
  argvals_template,
  argvals_target,
  lambda,
  min_span
)
}
\description{
Elastic partial matching of a template within a longer target curve
}
