% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_peak_persistence}
\alias{alignment_peak_persistence}
\title{Peak persistence diagram across a sweep of lambda values}
\usage{
alignment_peak_persistence(data, argvals, lambdas, max_iter, tol)
}
\description{
Peak persistence diagram across a sweep of lambda values
}
