% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_phase_boxplot_rust}
\alias{alignment_phase_boxplot_rust}
\title{Phase boxplot for warping functions}
\usage{
alignment_phase_boxplot_rust(gammas, argvals, factor)
}
\description{
Phase boxplot for warping functions
}
