% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_quality_compute}
\alias{alignment_quality_compute}
\title{Compute alignment quality metrics}
\usage{
alignment_quality_compute(
  data,
  mean,
  mean_srsf,
  gammas,
  aligned_data,
  argvals,
  n_iter,
  converged
)
}
\description{
Compute alignment quality metrics
}
