% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_reparameterize}
\alias{alignment_reparameterize}
\title{Apply warping function to reparameterize a curve}
\usage{
alignment_reparameterize(f, argvals, gamma)
}
\description{
Apply warping function to reparameterize a curve
}
