% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_robust_karcher_mean}
\alias{alignment_robust_karcher_mean}
\title{Robust (trimmed) Karcher mean in elastic metric}
\usage{
alignment_robust_karcher_mean(
  data,
  argvals,
  max_iter,
  tol,
  lambda,
  trim_fraction
)
}
\description{
Robust (trimmed) Karcher mean in elastic metric
}
