% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_shape_ci}
\alias{alignment_shape_ci}
\title{Bootstrap shape confidence intervals for the elastic Karcher mean}
\usage{
alignment_shape_ci(
  data,
  argvals,
  n_bootstrap,
  confidence_level,
  lambda,
  max_iter,
  tol,
  seed
)
}
\description{
Bootstrap shape confidence intervals for the elastic Karcher mean
}
