% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_shape_distance_rust}
\alias{alignment_shape_distance_rust}
\title{Elastic shape distance between two curves}
\usage{
alignment_shape_distance_rust(f1, f2, argvals, quotient, lambda)
}
\description{
Elastic shape distance between two curves
}
