% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_shape_mean_rust}
\alias{alignment_shape_mean_rust}
\title{Shape mean (Karcher mean in quotient space)}
\usage{
alignment_shape_mean_rust(data, argvals, quotient, lambda, max_iter, tol)
}
\description{
Shape mean (Karcher mean in quotient space)
}
