% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_shape_self_distance_matrix_rust}
\alias{alignment_shape_self_distance_matrix_rust}
\title{Shape self-distance matrix}
\usage{
alignment_shape_self_distance_matrix_rust(data, argvals, quotient, lambda)
}
\description{
Shape self-distance matrix
}
