% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_srsf_inverse}
\alias{alignment_srsf_inverse}
\title{Inverse SRSF: reconstruct curve from SRSF representation}
\usage{
alignment_srsf_inverse(q, argvals, f0)
}
\description{
Inverse SRSF: reconstruct curve from SRSF representation
}
