% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_srsf_transform}
\alias{alignment_srsf_transform}
\title{SRSF transform of functional data}
\usage{
alignment_srsf_transform(data, argvals)
}
\description{
SRSF transform of functional data
}
