% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_transfer}
\alias{alignment_transfer}
\title{Transfer alignment: align curves across populations}
\usage{
alignment_transfer(source_data, target_data, argvals, lambda, max_iter, tol)
}
\description{
Transfer alignment: align curves across populations
}
