% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_tsrvf_from_karcher}
\alias{alignment_tsrvf_from_karcher}
\title{Compute TSRVF from a pre-computed Karcher mean}
\usage{
alignment_tsrvf_from_karcher(
  mean,
  mean_srsf,
  gammas,
  aligned_data,
  argvals,
  n_iter,
  converged
)
}
\description{
Compute TSRVF from a pre-computed Karcher mean
}
