% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_tsrvf_inverse}
\alias{alignment_tsrvf_inverse}
\title{Inverse TSRVF: reconstruct curves from tangent vectors}
\usage{
alignment_tsrvf_inverse(
  tangent_vectors,
  `_mean`,
  mean_srsf,
  mean_srsf_norm,
  srsf_norms,
  `_gammas`,
  argvals,
  `_converged`
)
}
\description{
Inverse TSRVF: reconstruct curves from tangent vectors
}
