% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_warp_inverse_error_rust}
\alias{alignment_warp_inverse_error_rust}
\title{Warp inverse error: max |gamma(gamma_inv(t)) - t|}
\usage{
alignment_warp_inverse_error_rust(gamma, argvals)
}
\description{
Warp inverse error: max |gamma(gamma_inv(t)) - t|
}
