% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_warp_statistics_rust}
\alias{alignment_warp_statistics_rust}
\title{Warp statistics: mean, variance, confidence bands, Karcher mean warp}
\usage{
alignment_warp_statistics_rust(gammas, argvals, confidence_level)
}
\description{
Warp statistics: mean, variance, confidence bands, Karcher mean warp
}
