% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{alignment_with_landmarks}
\alias{alignment_with_landmarks}
\title{Elastic alignment with automatic landmark detection}
\usage{
alignment_with_landmarks(
  f1,
  f2,
  argvals,
  kind,
  min_prominence,
  expected_count,
  lambda
)
}
\description{
Elastic alignment with automatic landmark detection
}
