% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{basis2fdata_2d_rust}
\alias{basis2fdata_2d_rust}
\title{Reconstruct 2D functional data from tensor product basis coefficients (low-level)}
\usage{
basis2fdata_2d_rust(
  coefs,
  argvals_s,
  argvals_t,
  nbasis_s,
  nbasis_t,
  basis_type
)
}
\description{
Reconstruct 2D functional data from tensor product basis coefficients (low-level)
}
