% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bootstrap_ci_fregre_lm_rust}
\alias{bootstrap_ci_fregre_lm_rust}
\title{Bootstrap confidence intervals for β(t) in functional linear model}
\usage{
bootstrap_ci_fregre_lm_rust(
  data,
  y,
  scalar_covariates,
  ncomp,
  n_boot,
  alpha,
  seed
)
}
\description{
Bootstrap confidence intervals for β(t) in functional linear model
}
