% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calinski_harabasz}
\alias{calinski_harabasz}
\title{Compute Calinski-Harabasz index (variance ratio criterion)
Higher values indicate better defined clusters}
\usage{
calinski_harabasz(data, clusters)
}
\description{
Compute Calinski-Harabasz index (variance ratio criterion)
Higher values indicate better defined clusters
}
