% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{conformal.elastic.pcr}
\alias{conformal.elastic.pcr}
\title{Conformal Prediction for Elastic PCR}
\usage{
conformal.elastic.pcr(
  fdataobj,
  y,
  newdata,
  ncomp = 3,
  pca.method = c("vertical", "horizontal", "joint"),
  lambda = 0,
  cal.fraction = 0.25,
  alpha = 0.1,
  seed = NULL
)
}
\arguments{
\item{fdataobj}{An object of class 'fdata' (training data).}

\item{y}{Response vector (training).}

\item{newdata}{An object of class 'fdata' (test data).}

\item{ncomp}{Number of FPC components (default 3).}

\item{pca.method}{PCA method: "vertical", "horizontal", or "joint".}

\item{lambda}{Regularization (default 0).}

\item{cal.fraction}{Calibration fraction (default 0.25).}

\item{alpha}{Miscoverage level (default 0.1).}

\item{seed}{Random seed.}
}
\value{
Same as \code{\link{conformal.fregre.lm}}.
}
\description{
Conformal Prediction for Elastic PCR
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
y <- rnorm(50)
cp <- conformal.elastic.pcr(fd[1:40, ], y[1:40], fd[41:50, ])
}

}
