% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{conformal.fregre.np}
\alias{conformal.fregre.np}
\title{Conformal Prediction for Nonparametric Functional Regression}
\usage{
conformal.fregre.np(
  fdataobj,
  y,
  newdata,
  scalar.train = NULL,
  scalar.test = NULL,
  h.func = 0,
  h.scalar = 0,
  cal.fraction = 0.25,
  alpha = 0.1,
  seed = NULL
)
}
\arguments{
\item{fdataobj}{An object of class 'fdata' (training data).}

\item{y}{Response vector (training).}

\item{newdata}{An object of class 'fdata' (test data).}

\item{scalar.train}{Optional scalar covariates for training.}

\item{scalar.test}{Optional scalar covariates for test.}

\item{h.func}{Functional bandwidth (0 for auto).}

\item{h.scalar}{Scalar bandwidth (0 for auto).}

\item{cal.fraction}{Fraction of data for calibration (default 0.25).}

\item{alpha}{Miscoverage level (default 0.1).}

\item{seed}{Random seed.}
}
\value{
Same as \code{\link{conformal.fregre.lm}}.
}
\description{
Conformal Prediction for Nonparametric Functional Regression
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
y <- rnorm(50)
cp <- conformal.fregre.np(fd[1:40, ], y[1:40], fd[41:50, ])
}

}
