% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_classif_rust}
\alias{conformal_classif_rust}
\title{Conformal classification}
\usage{
conformal_classif_rust(
  data,
  y,
  test_data,
  covariates_train,
  covariates_test,
  ncomp,
  classifier,
  k_nn,
  score_type,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Conformal classification
}
