% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_elastic_logistic_rust}
\alias{conformal_elastic_logistic_rust}
\title{Conformal prediction for elastic logistic regression}
\usage{
conformal_elastic_logistic_rust(
  data,
  y,
  test_data,
  argvals,
  lambda,
  score_type,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Conformal prediction for elastic logistic regression
}
