% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_elastic_pcr_rust}
\alias{conformal_elastic_pcr_rust}
\title{Conformal prediction for elastic PCR}
\usage{
conformal_elastic_pcr_rust(
  data,
  y,
  test_data,
  argvals,
  ncomp,
  pca_method,
  lambda,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Conformal prediction for elastic PCR
}
