% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_elastic_regression_rust}
\alias{conformal_elastic_regression_rust}
\title{Conformal prediction for elastic regression}
\usage{
conformal_elastic_regression_rust(
  data,
  y,
  test_data,
  argvals,
  ncomp_beta,
  lambda,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Conformal prediction for elastic regression
}
