% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_fregre_lm_rust}
\alias{conformal_fregre_lm_rust}
\title{Conformal prediction for functional linear model}
\usage{
conformal_fregre_lm_rust(
  data,
  y,
  test_data,
  scalar_train,
  scalar_test,
  ncomp,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Conformal prediction for functional linear model
}
