% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_fregre_np_rust}
\alias{conformal_fregre_np_rust}
\title{Conformal prediction for nonparametric functional regression}
\usage{
conformal_fregre_np_rust(
  data,
  y,
  test_data,
  argvals,
  scalar_train,
  scalar_test,
  h_func,
  h_scalar,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Conformal prediction for nonparametric functional regression
}
