% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_generic_classification_logistic_rust}
\alias{conformal_generic_classification_logistic_rust}
\title{Generic conformal classification for a pre-fitted logistic model}
\usage{
conformal_generic_classification_logistic_rust(
  data,
  y,
  test_data,
  scalar_train,
  scalar_test,
  fpca_mean,
  fpca_rotation,
  fpca_scores,
  ncomp,
  intercept,
  coefficients,
  gamma,
  score_type,
  calibration_indices,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Generic conformal classification for a pre-fitted logistic model
}
