% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_generic_regression_lm_rust}
\alias{conformal_generic_regression_lm_rust}
\title{Generic conformal regression for a pre-fitted fregre.lm model}
\usage{
conformal_generic_regression_lm_rust(
  data,
  y,
  test_data,
  scalar_train,
  scalar_test,
  fpca_mean,
  fpca_rotation,
  fpca_scores,
  ncomp,
  coefficients,
  gamma,
  calibration_indices,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Generic conformal regression for a pre-fitted fregre.lm model
}
