% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{conformal_logistic_rust}
\alias{conformal_logistic_rust}
\title{Conformal prediction for logistic regression}
\usage{
conformal_logistic_rust(
  data,
  y,
  test_data,
  scalar_train,
  scalar_test,
  ncomp,
  max_iter,
  tol,
  score_type,
  cal_fraction,
  alpha,
  seed
)
}
\description{
Conformal prediction for logistic regression
}
