% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{curve.geodesic}
\alias{curve.geodesic}
\title{Geodesic Path Between Two Curves}
\usage{
curve.geodesic(f1, f2, argvals, n.points = 10, lambda = 0)
}
\arguments{
\item{f1}{Numeric vector of the first curve's values.}

\item{f2}{Numeric vector of the second curve's values.}

\item{argvals}{Numeric vector of the evaluation grid (common to both curves).}

\item{n.points}{Number of points along the geodesic to compute (default 10).}

\item{lambda}{Regularisation parameter controlling warping smoothness
(default 0).}
}
\value{
An object of class 'curve.geodesic' with components:
\describe{
\item{path}{fdata of the interpolated curves along the geodesic}
\item{distance}{the total geodesic (elastic) distance between f1 and f2}
\item{gamma}{numeric vector of the optimal warping function}
\item{argvals}{the evaluation grid}
}
}
\description{
Compute the geodesic (shortest path) between two curves in the elastic
metric. The geodesic is represented as a sequence of intermediate curves
interpolating between the two endpoints in the SRVF space.
}
\examples{
\donttest{
t <- seq(0, 1, length.out = 100)
f1 <- sin(2 * pi * t)
f2 <- cos(2 * pi * t)
geo <- curve.geodesic(f1, f2, t, n.points = 8)
}
}
\references{
Srivastava, A. and Klassen, E. (2016). \emph{Functional and Shape Data
Analysis}. Springer.
}
\seealso{
\code{\link{elastic.distance}}
}
