% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cv_conformal_classif_rust}
\alias{cv_conformal_classif_rust}
\title{CV-conformal classification}
\usage{
cv_conformal_classif_rust(
  data,
  y,
  test_data,
  covariates_train,
  covariates_test,
  ncomp,
  classifier,
  k_nn,
  score_type,
  n_folds,
  alpha,
  seed
)
}
\description{
CV-conformal classification
}
