% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cv_conformal_fregre_lm_rust}
\alias{cv_conformal_fregre_lm_rust}
\title{CV-conformal regression using functional linear model}
\usage{
cv_conformal_fregre_lm_rust(
  data,
  y,
  test_data,
  scalar_train,
  scalar_test,
  ncomp,
  n_folds,
  alpha,
  seed
)
}
\description{
CV-conformal regression using functional linear model
}
