% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cv_conformal_fregre_np_rust}
\alias{cv_conformal_fregre_np_rust}
\title{CV-conformal regression using nonparametric kernel regression}
\usage{
cv_conformal_fregre_np_rust(
  data,
  y,
  test_data,
  argvals,
  scalar_train,
  scalar_test,
  h_func,
  h_scalar,
  n_folds,
  alpha,
  seed
)
}
\description{
CV-conformal regression using nonparametric kernel regression
}
