% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_bd_1d}
\alias{depth_bd_1d}
\title{Band Depth (BD) for 1D functional data
BD(x) = proportion of pairs (i,j) where x lies within the band formed by curves i and j
A curve lies in the band if at every time point t, min(X_i(t), X_j(t)) <= x(t) <= max(X_i(t), X_j(t))}
\usage{
depth_bd_1d(fdataobj, fdataori)
}
\description{
Band Depth (BD) for 1D functional data
BD(x) = proportion of pairs (i,j) where x lies within the band formed by curves i and j
A curve lies in the band if at every time point t, min(X_i(t), X_j(t)) <= x(t) <= max(X_i(t), X_j(t))
}
