% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_fm_1d}
\alias{depth_fm_1d}
\title{Compute Fraiman-Muniz depth}
\usage{
depth_fm_1d(fdataobj, fdataori, `_trim`, scale)
}
\description{
Uses the FM1 formula from fda.usc: d = 1 - |0.5 - Fn(x)|
With scale=TRUE (default): d = (1 - |0.5 - Fn(x)| - 0.5) * 2 = 2 * min(Fn(x), 1-Fn(x))
}
