% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_fm_2d}
\alias{depth_fm_2d}
\title{Fraiman-Muniz depth for 2D functional data (surfaces)
Integrates univariate depth over (s,t) grid}
\usage{
depth_fm_2d(fdataobj, fdataori, m1, m2, scale)
}
\description{
Fraiman-Muniz depth for 2D functional data (surfaces)
Integrates univariate depth over (s,t) grid
}
