% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_fsd_2d}
\alias{depth_fsd_2d}
\title{Functional Spatial Depth for 2D functional data}
\usage{
depth_fsd_2d(fdataobj, fdataori, m1, m2)
}
\description{
Functional Spatial Depth for 2D functional data
}
