% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_kfsd_2d}
\alias{depth_kfsd_2d}
\title{Kernel Functional Spatial Depth (KFSD) for 2D functional data
Implements the RKHS-based formulation matching fda.usc}
\usage{
depth_kfsd_2d(fdataobj, fdataori, `_m1`, `_m2`, h)
}
\description{
Kernel Functional Spatial Depth (KFSD) for 2D functional data
Implements the RKHS-based formulation matching fda.usc
}
