% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_mbd_1d}
\alias{depth_mbd_1d}
\title{Modified Band Depth (MBD) for 1D functional data
MBD(x) = average over pairs (i,j) of the proportion of the domain where x is inside the band
This is more robust than BD as it doesn't require complete containment}
\usage{
depth_mbd_1d(fdataobj, fdataori)
}
\description{
Modified Band Depth (MBD) for 1D functional data
MBD(x) = average over pairs (i,j) of the proportion of the domain where x is inside the band
This is more robust than BD as it doesn't require complete containment
}
