% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_mei_1d}
\alias{depth_mei_1d}
\title{Modified Epigraph Index (MEI) for 1D functional data
MEI measures the proportion of time a curve is below other curves
MEI(x_i) = (1/n) * sum_j (1/m) * sum_t I(x_i(t) < x_j(t)) + 0.5*I(x_i(t) = x_j(t))}
\usage{
depth_mei_1d(fdataobj, fdataori)
}
\description{
Modified Epigraph Index (MEI) for 1D functional data
MEI measures the proportion of time a curve is below other curves
MEI(x_i) = (1/n) * sum_j (1/m) * sum_t I(x_i(t) < x_j(t)) + 0.5*I(x_i(t) = x_j(t))
}
