% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_mode_2d}
\alias{depth_mode_2d}
\title{Modal depth for 2D functional data (surfaces)
Uses L2 distance in the flattened surface space}
\usage{
depth_mode_2d(fdataobj, fdataori, m1, m2, h)
}
\description{
Modal depth for 2D functional data (surfaces)
Uses L2 distance in the flattened surface space
}
