% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_rp_1d_seeded}
\alias{depth_rp_1d_seeded}
\title{Random projection depth with optional seed}
\usage{
depth_rp_1d_seeded(data_obj, data_ori, nproj, seed)
}
\description{
Random projection depth with optional seed
}
