% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_rp_2d}
\alias{depth_rp_2d}
\title{Random projection depth for 2D functional data (surfaces)
Projects surfaces to scalars using random projections}
\usage{
depth_rp_2d(fdataobj, fdataori, m1, m2, nproj)
}
\description{
Random projection depth for 2D functional data (surfaces)
Projects surfaces to scalars using random projections
}
