% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_rpd_1d_rust}
\alias{depth_rpd_1d_rust}
\title{Random Projection Depth with Derivatives (seeded)}
\usage{
depth_rpd_1d_rust(data_obj, data_ori, argvals, nproj, nderiv, seed)
}
\description{
Random Projection Depth with Derivatives (seeded)
}
