% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_rt_1d_seeded}
\alias{depth_rt_1d_seeded}
\title{Random Tukey depth with optional seed}
\usage{
depth_rt_1d_seeded(data_obj, data_ori, nproj, seed)
}
\description{
Random Tukey depth with optional seed
}
