% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{depth_rt_2d}
\alias{depth_rt_2d}
\title{Random Tukey depth for 2D functional data (surfaces)}
\usage{
depth_rt_2d(fdataobj, fdataori, m1, m2, nproj)
}
\description{
Random Tukey depth for 2D functional data (surfaces)
}
