% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{.compute_fold_metrics}
\alias{.compute_fold_metrics}
\title{Compute Per-Fold CV Metrics}
\usage{
.compute_fold_metrics(y, oof, folds, kfold, type)
}
\description{
Compute Per-Fold CV Metrics
}
\keyword{internal}
