% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{.compute_metrics_summary}
\alias{.compute_metrics_summary}
\title{Compute Metrics Summary Across Repetitions}
\usage{
.compute_metrics_summary(rep_metrics, type)
}
\description{
Compute Metrics Summary Across Repetitions
}
\keyword{internal}
