% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{.compute_rep_metrics}
\alias{.compute_rep_metrics}
\title{Compute Per-Repetition Metrics}
\usage{
.compute_rep_metrics(y, oof_mat, nrep, type)
}
\description{
Compute Per-Repetition Metrics
}
\keyword{internal}
