% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{.create_folds}
\alias{.create_folds}
\title{Create Stratified or Random Fold Assignments}
\usage{
.create_folds(y, kfold, type, stratified, seed)
}
\arguments{
\item{y}{Response vector.}

\item{kfold}{Number of folds.}

\item{type}{\code{"regression"} or \code{"classification"}.}

\item{stratified}{Logical.}

\item{seed}{Optional seed.}
}
\value{
Integer vector of fold assignments.
}
\description{
Create Stratified or Random Fold Assignments
}
\keyword{internal}
