% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{eigenfunctions_1d}
\alias{eigenfunctions_1d}
\title{Compute eigenfunction basis values
efun_type: 0 = Fourier, 1 = Poly, 2 = PolyHigh, 3 = Wiener}
\usage{
eigenfunctions_1d(argvals, m, efun_type)
}
\description{
Compute eigenfunction basis values
efun_type: 0 = Fourier, 1 = Poly, 2 = PolyHigh, 3 = Wiener
}
