% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{eigenvalues_1d}
\alias{eigenvalues_1d}
\title{Generate eigenvalue sequence
eval_type: 0 = linear, 1 = exponential, 2 = wiener}
\usage{
eigenvalues_1d(m, eval_type)
}
\description{
Generate eigenvalue sequence
eval_type: 0 = linear, 1 = exponential, 2 = wiener
}
