% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_clustering.R
\name{elastic.cutree}
\alias{elastic.cutree}
\title{Cut Elastic Dendrogram}
\usage{
elastic.cutree(tree, k)
}
\arguments{
\item{tree}{An object of class \code{elastic.hclust}.}

\item{k}{Integer: desired number of clusters.}
}
\value{
An integer vector of cluster labels (1-indexed), one per curve.
}
\description{
Cut an elastic hierarchical clustering dendrogram to produce a specified
number of clusters.
}
\examples{
\donttest{
set.seed(1)
t <- seq(0, 1, length.out = 30)
X <- matrix(0, 10, 30)
for (i in 1:5)  X[i, ] <- sin(2 * pi * t) + rnorm(30, 0, 0.1)
for (i in 6:10) X[i, ] <- cos(2 * pi * t) + rnorm(30, 0, 0.1)
fd <- fdata(X, argvals = t)
hc <- elastic.hclust(fd, method = "complete")
labels <- elastic.cutree(hc, k = 2)
table(labels)
}
}
\seealso{
\code{\link{elastic.hclust}}
}
