% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_clustering.R
\name{elastic.hclust}
\alias{elastic.hclust}
\title{Elastic Hierarchical Clustering}
\usage{
elastic.hclust(
  fdataobj,
  method = c("complete", "single", "average"),
  lambda = 0
)
}
\arguments{
\item{fdataobj}{An object of class \code{fdata}.}

\item{method}{Character: linkage method. One of \code{"complete"}
(default), \code{"single"}, or \code{"average"}.}

\item{lambda}{Regularization parameter for elastic alignment (default 0).}
}
\value{
An object of class \code{elastic.hclust} with components:
\describe{
\item{merges}{A 3-column matrix of merge steps (i, j, distance),
with 1-indexed cluster labels}
\item{distance.matrix}{The pairwise elastic distance matrix}
\item{method}{The linkage method used}
\item{fdataobj}{Original fdata object}
\item{call}{The matched call}
}
}
\description{
Perform agglomerative hierarchical clustering on functional data using
pairwise elastic (Fisher-Rao) distances. Supports single, complete,
and average linkage.
}
\examples{
\donttest{
set.seed(1)
t <- seq(0, 1, length.out = 30)
X <- matrix(0, 10, 30)
for (i in 1:5)  X[i, ] <- sin(2 * pi * t) + rnorm(30, 0, 0.1)
for (i in 6:10) X[i, ] <- cos(2 * pi * t) + rnorm(30, 0, 0.1)
fd <- fdata(X, argvals = t)
hc <- elastic.hclust(fd, method = "complete")
hc
}
}
\references{
Srivastava, A. and Klassen, E. (2016). \emph{Functional and Shape Data
Analysis}. Springer.
}
\seealso{
\code{\link{elastic.cutree}} for cutting the dendrogram,
\code{\link{elastic.kmeans}} for k-means clustering,
\code{\link{elastic.distance}} for the distance metric
}
