% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_cut_dendrogram_rust}
\alias{elastic_cut_dendrogram_rust}
\title{Cut a hierarchical dendrogram to get cluster labels}
\usage{
elastic_cut_dendrogram_rust(merges_data, distance_matrix, k)
}
\description{
Cut a hierarchical dendrogram to get cluster labels
}
