% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_fpca_changepoint_rust}
\alias{elastic_fpca_changepoint_rust}
\title{FPCA-based changepoint detection}
\usage{
elastic_fpca_changepoint_rust(
  data,
  argvals,
  pca_method,
  ncomp,
  lambda,
  max_iter,
  n_mc,
  seed
)
}
\description{
FPCA-based changepoint detection
}
