% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_hierarchical_rust}
\alias{elastic_hierarchical_rust}
\title{Elastic hierarchical clustering}
\usage{
elastic_hierarchical_rust(data, argvals, method, lambda)
}
\description{
Elastic hierarchical clustering
}
