% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_horiz_fpca_rust}
\alias{elastic_horiz_fpca_rust}
\title{Horizontal FPCA on warping functions}
\usage{
elastic_horiz_fpca_rust(
  aligned_data,
  gammas_data,
  mean_curve,
  mean_srsf,
  aligned_srsfs,
  argvals,
  ncomp
)
}
\description{
Horizontal FPCA on warping functions
}
