% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_joint_fpca_rust}
\alias{elastic_joint_fpca_rust}
\title{Joint (amplitude + phase) FPCA after elastic alignment}
\usage{
elastic_joint_fpca_rust(
  aligned_data,
  gammas,
  mean_data,
  mean_srsf,
  aligned_srsfs,
  argvals,
  ncomp
)
}
\description{
Joint (amplitude + phase) FPCA after elastic alignment
}
