% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_pcr_attribution_rust}
\alias{elastic_pcr_attribution_rust}
\title{Elastic PCR amplitude vs phase attribution}
\usage{
elastic_pcr_attribution_rust(
  data,
  y,
  argvals,
  ncomp,
  pca_method,
  lambda,
  max_iter,
  tol,
  n_perm,
  seed
)
}
\description{
Elastic PCR amplitude vs phase attribution
}
