% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_pcr_rust}
\alias{elastic_pcr_rust}
\title{Elastic principal component regression}
\usage{
elastic_pcr_rust(data, y, argvals, ncomp, pca_method, lambda, max_iter, tol)
}
\description{
Elastic principal component regression
}
