% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_ph_changepoint_rust}
\alias{elastic_ph_changepoint_rust}
\title{Phase changepoint detection}
\usage{
elastic_ph_changepoint_rust(
  data,
  argvals,
  lambda,
  max_iter,
  n_mc,
  cov_kernel,
  cov_bandwidth,
  seed
)
}
\description{
Phase changepoint detection
}
