% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_regression_rust}
\alias{elastic_regression_rust}
\title{Elastic scalar-on-function regression}
\usage{
elastic_regression_rust(data, y, argvals, ncomp_beta, lambda, max_iter, tol)
}
\description{
Elastic scalar-on-function regression
}
