% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_spm_phase1_rust}
\alias{elastic_spm_phase1_rust}
\title{Elastic SPM Phase I.}
\usage{
elastic_spm_phase1_rust(
  data,
  argvals,
  ncomp,
  alpha,
  tuning_fraction,
  seed,
  align_lambda,
  monitor_phase,
  warp_ncomp
)
}
\description{
Elastic SPM Phase I.
}
