% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_vert_fpca_rust}
\alias{elastic_vert_fpca_rust}
\title{Vertical FPCA on elastic-aligned data}
\usage{
elastic_vert_fpca_rust(
  aligned_data,
  gammas_data,
  mean_curve,
  mean_srsf,
  aligned_srsfs,
  argvals,
  ncomp
)
}
\description{
Vertical FPCA on elastic-aligned data
}
