% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{explain_importance_rust}
\alias{explain_importance_rust}
\title{FPC permutation importance}
\usage{
explain_importance_rust(
  intercept,
  beta_t,
  beta_se,
  gamma,
  fitted_values,
  residuals,
  r_squared,
  r_squared_adj,
  std_errors,
  ncomp,
  coefficients,
  residual_se,
  gcv,
  fpca_mean,
  fpca_rotation_data,
  fpca_rotation_nrow,
  fpca_rotation_ncol,
  fpca_scores_data,
  fpca_scores_nrow,
  fpca_scores_ncol,
  data,
  y,
  n_perm,
  seed
)
}
\description{
FPC permutation importance
}
