% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{explain_prototype_rust}
\alias{explain_prototype_rust}
\title{Prototype criticism}
\usage{
explain_prototype_rust(
  fpca_scores_data,
  fpca_scores_nrow,
  fpca_scores_ncol,
  ncomp,
  n_prototypes,
  n_criticisms
)
}
\description{
Prototype criticism
}
