% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{explain_saliency_logistic_rust}
\alias{explain_saliency_logistic_rust}
\title{Saliency map for logistic model (fit-only, no data)}
\usage{
explain_saliency_logistic_rust(
  intercept,
  beta_t,
  beta_se,
  gamma,
  probabilities,
  predicted_classes,
  ncomp,
  accuracy,
  std_errors,
  coefficients,
  log_likelihood,
  iterations,
  fpca_mean,
  fpca_rotation_data,
  fpca_rotation_nrow,
  fpca_rotation_ncol,
  fpca_scores_data,
  fpca_scores_nrow,
  fpca_scores_ncol
)
}
\description{
Saliency map for logistic model (fit-only, no data)
}
