% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{explain_significant_regions_rust}
\alias{explain_significant_regions_rust}
\title{Significant regions from standard errors}
\usage{
explain_significant_regions_rust(beta_t, beta_se, alpha)
}
\description{
Significant regions from standard errors
}
