% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{fanova}
\alias{fanova}
\title{Functional ANOVA}
\usage{
fanova(fdataobj, groups, n.perm = 1000)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{groups}{Integer vector of group labels.}

\item{n.perm}{Number of permutations (default 1000).}
}
\value{
An object of class 'fanova' with components:
\item{group.means}{Group mean functions as fdata}
\item{overall.mean}{Overall mean function}
\item{f.statistic.t}{Pointwise F-statistic}
\item{global.statistic}{Global test statistic}
\item{p.value}{P-value from permutation test}
}
\description{
Tests for differences in mean functions across groups using a
permutation-based F-test.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
groups <- rep(1:2, each = 25)
res <- fanova(fd, groups = groups, n.perm = 100)
res
}

}
