% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fclassif_cv_rust}
\alias{fclassif_cv_rust}
\title{Cross-validated classification}
\usage{
fclassif_cv_rust(data, argvals, y, covariates, method, ncomp, nfold, seed)
}
\description{
Cross-validated classification
}
