% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fclassif_kernel_rust}
\alias{fclassif_kernel_rust}
\title{Kernel classification}
\usage{
fclassif_kernel_rust(data, argvals, y, covariates, h_func, h_scalar)
}
\description{
Kernel classification
}
