% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdars-package.R
\docType{package}
\name{fdars-package}
\alias{fdars}
\alias{fdars-package}
\title{fdars: Functional Data Analysis in Rust}
\description{
Functional data analysis tools with a high-performance Rust backend. Provides methods for functional data manipulation, depth computation, distance metrics, regression, and statistical testing. Supports both 1D functional data (curves) and 2D functional data (surfaces).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sipemu.github.io/fdars/}
  \item \url{https://github.com/sipemu/fdars}
  \item Report bugs at \url{https://github.com/sipemu/fdars/issues}
}

}
\author{
\strong{Maintainer}: Simon Müller \email{simon.mueller@muon-stat.com}

}
\keyword{internal}
