% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{fdata.gradient}
\alias{fdata.gradient}
\title{High-Accuracy Gradient for Functional Data}
\usage{
fdata.gradient(fdataobj)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}
}
\value{
An \code{fdata} object with the gradient values.
}
\description{
Computes the numerical gradient (first derivative) of functional data using
a 5-point stencil for uniform grids (O(h^4) accuracy) or 3-point Lagrange
for non-uniform grids. Complementary to \code{deriv.fdata} which uses
central differences.
}
\examples{
t <- seq(0, 1, length.out = 100)
X <- matrix(sin(2 * pi * t), nrow = 1)
fd <- fdata(X, argvals = t)
gd <- fdata.gradient(fd)
}
