% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fdata2basis_1d}
\alias{fdata2basis_1d}
\title{Convert functional data to basis coefficients
type: 0 = bspline, 1 = fourier}
\usage{
fdata2basis_1d(data, argvals, nbasis, basis_type)
}
\description{
Convert functional data to basis coefficients
type: 0 = bspline, 1 = fourier
}
