% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fdata2basis_2d_rust}
\alias{fdata2basis_2d_rust}
\title{Project 2D functional data to tensor product basis coefficients (low-level)}
\usage{
fdata2basis_2d_rust(data, argvals_s, argvals_t, nbasis_s, nbasis_t, basis_type)
}
\description{
Project 2D functional data to tensor product basis coefficients (low-level)
}
