% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fdata2pc_1d}
\alias{fdata2pc_1d}
\title{Perform functional PCA via SVD on centered data
Returns: singular values, rotation matrix (loadings), scores, mean}
\usage{
fdata2pc_1d(data, ncomp, `_lambda`)
}
\description{
Perform functional PCA via SVD on centered data
Returns: singular values, rotation matrix (loadings), scores, mean
}
