% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fdata2pls_1d}
\alias{fdata2pls_1d}
\title{Perform PLS via NIPALS algorithm
Returns: weights, scores, loadings}
\usage{
fdata2pls_1d(data, y, ncomp, `_lambda`)
}
\description{
Perform PLS via NIPALS algorithm
Returns: weights, scores, loadings
}
