% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fdata_center_1d}
\alias{fdata_center_1d}
\title{Center functional data by subtracting the mean function}
\usage{
fdata_center_1d(data)
}
\description{
Center functional data by subtracting the mean function
}
