% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fdata_deriv_1d}
\alias{fdata_deriv_1d}
\title{Compute numerical derivative of functional data (parallelized over rows)}
\usage{
fdata_deriv_1d(data, argvals, nderiv)
}
\description{
Compute numerical derivative of functional data (parallelized over rows)
}
