% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fdata_gradient_rust}
\alias{fdata_gradient_rust}
\title{High-accuracy gradient using 5-point stencil (uniform) or 3-point Lagrange (non-uniform)}
\usage{
fdata_gradient_rust(data, argvals)
}
\description{
High-accuracy gradient using 5-point stencil (uniform) or 3-point Lagrange (non-uniform)
}
