% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fdata_mean_2d}
\alias{fdata_mean_2d}
\title{Compute the mean function across all samples (2D surfaces)
Data is stored as n x (m1*m2) matrix where each row is a flattened surface}
\usage{
fdata_mean_2d(data)
}
\description{
Compute the mean function across all samples (2D surfaces)
Data is stored as n x (m1*m2) matrix where each row is a flattened surface
}
