% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmm.R
\name{fmm.predict}
\alias{fmm.predict}
\title{Predict from Functional Mixed Model}
\usage{
fmm.predict(object, new.covariates = NULL, ...)
}
\arguments{
\item{object}{A fitted 'fmm' object.}

\item{new.covariates}{Matrix of new covariate values. If NULL, returns
the overall mean function.}

\item{...}{Additional arguments (ignored).}
}
\value{
An fdata object of predicted functional values.
}
\description{
Predict population-level (fixed effects only) functional responses
for new covariate values.
}
