% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fmm_predict_rust}
\alias{fmm_predict_rust}
\title{Predict from functional mixed model}
\usage{
fmm_predict_rust(mean_function, beta_data, new_covariates)
}
\description{
Predict from functional mixed model
}
