% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fmm_test_fixed_rust}
\alias{fmm_test_fixed_rust}
\title{Permutation test for fixed effects in FMM}
\usage{
fmm_test_fixed_rust(data, subject_ids, covariates, ncomp, n_perm, seed)
}
\description{
Permutation test for fixed effects in FMM
}
