% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{fosr.fpc}
\alias{fosr.fpc}
\title{FPC-based Function-on-Scalar Regression}
\usage{
fosr.fpc(fdataobj, predictors, ncomp = 3)
}
\arguments{
\item{fdataobj}{An object of class 'fdata' (functional response).}

\item{predictors}{A matrix of scalar predictors (n x p).}

\item{ncomp}{Number of FPC components (default 3).}
}
\value{
An object of class 'fosr'.
}
\description{
Fits a function-on-scalar regression using functional principal components.
}
\examples{
\donttest{
Y <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
X <- cbind(rnorm(50), rnorm(50))
fit <- fosr.fpc(Y, predictors = X, ncomp = 2)
fit
}

}
