% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fosr_2d_rust}
\alias{fosr_2d_rust}
\title{2D Function-on-scalar regression with tensor-product penalties}
\usage{
fosr_2d_rust(data, predictors, argvals_s, argvals_t, lambda_s, lambda_t)
}
\description{
2D Function-on-scalar regression with tensor-product penalties
}
