% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{frcc_phase1_rust}
\alias{frcc_phase1_rust}
\title{FRCC Phase I: Build a functional regression control chart.}
\usage{
frcc_phase1_rust(
  y_curves,
  predictors,
  argvals,
  ncomp,
  fosr_lambda,
  alpha,
  tuning_fraction,
  seed
)
}
\description{
FRCC Phase I: Build a functional regression control chart.
}
