% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{fregre.l1}
\alias{fregre.l1}
\title{L1 (Least Absolute Deviation) Functional Regression}
\usage{
fregre.l1(fdataobj, y, scalar.covariates = NULL, ncomp = 3)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{y}{Response vector.}

\item{scalar.covariates}{Optional matrix of scalar covariates.}

\item{ncomp}{Number of FPC components (default 3).}
}
\value{
An object of class 'fregre.robust' with components:
\describe{
\item{intercept}{Intercept}
\item{beta.t}{Functional coefficient beta(t)}
\item{fitted.values}{Fitted response values}
\item{residuals}{Residuals}
\item{coefficients}{Regression coefficients}
\item{ncomp}{Number of FPC components used}
\item{iterations}{Number of IRLS iterations}
\item{converged}{Whether the algorithm converged}
\item{weights}{Final IRLS weights}
\item{r.squared}{R-squared statistic}
\item{method}{Either "l1" or "huber"}
}
}
\description{
Fits a functional linear model using L1 loss (median regression),
which is robust to outliers in the response.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
y <- rnorm(50)
fit <- fregre.l1(fd, y, ncomp = 3)
}

}
