% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fregre_cv_rust}
\alias{fregre_cv_rust}
\title{Cross-validation for FPC component selection}
\usage{
fregre_cv_rust(data, y, scalar_covariates, k_min, k_max, n_folds)
}
\description{
Cross-validation for FPC component selection
}
