% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fregre_huber_rust}
\alias{fregre_huber_rust}
\title{Huber M-estimation functional regression}
\usage{
fregre_huber_rust(data, y, scalar_covariates, ncomp, k)
}
\description{
Huber M-estimation functional regression
}
