% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fregre_np_mixed_rust}
\alias{fregre_np_mixed_rust}
\title{Nonparametric functional regression with mixed predictors}
\usage{
fregre_np_mixed_rust(data, y, argvals, scalar_covariates, h_func, h_scalar)
}
\description{
Nonparametric functional regression with mixed predictors
}
