% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{functional_logistic_rust}
\alias{functional_logistic_rust}
\title{Functional logistic regression}
\usage{
functional_logistic_rust(data, y, scalar_covariates, ncomp, max_iter, tol)
}
\description{
Functional logistic regression
}
