% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{gmm_cluster_rust}
\alias{gmm_cluster_rust}
\title{GMM clustering with automatic K selection}
\usage{
gmm_cluster_rust(
  data,
  argvals,
  covariates,
  k_range,
  nbasis,
  basis_type,
  cov_type,
  cov_weight,
  max_iter,
  tol,
  n_init,
  seed,
  use_icl
)
}
\description{
GMM clustering with automatic K selection
}
