% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{inprod_fdata}
\alias{inprod_fdata}
\title{Inner product of two functional data objects
<f, g> = integral(f(t) * g(t) dt)}
\usage{
inprod_fdata(data1, data2, argvals)
}
\description{
Inner product of two functional data objects
<f, g> = integral(f(t) * g(t) dt)
}
