% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{int_simpson}
\alias{int_simpson}
\title{Simpson's rule integration for functional data
Integrates each curve over the domain}
\usage{
int_simpson(data, argvals)
}
\description{
Simpson's rule integration for functional data
Integrates each curve over the domain
}
