% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{irreg_integrate}
\alias{irreg_integrate}
\title{Compute integral for each curve in irregular functional data}
\usage{
irreg_integrate(offsets, argvals, values)
}
\description{
Compute integral for each curve in irregular functional data
}
