% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{irreg_mean_kernel}
\alias{irreg_mean_kernel}
\title{Estimate mean function for irregular data using kernel smoothing}
\usage{
irreg_mean_kernel(
  offsets,
  argvals,
  values,
  target_argvals,
  bandwidth,
  kernel_type
)
}
\description{
Estimate mean function for irregular data using kernel smoothing
}
