% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{irreg_metric_lp}
\alias{irreg_metric_lp}
\title{Compute pairwise Lp distances for irregular functional data}
\usage{
irreg_metric_lp(offsets, argvals, values, p)
}
\description{
Compute pairwise Lp distances for irregular functional data
}
