% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{irreg_norm_lp}
\alias{irreg_norm_lp}
\title{Compute Lp norm for each curve in irregular functional data}
\usage{
irreg_norm_lp(offsets, argvals, values, p)
}
\description{
Compute Lp norm for each curve in irregular functional data
}
