% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{irreg_to_regular}
\alias{irreg_to_regular}
\title{Convert irregular data to regular grid via interpolation}
\usage{
irreg_to_regular(offsets, argvals, values, target_grid)
}
\description{
Convert irregular data to regular grid via interpolation
}
