% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{jackknife_plus_fregre_lm_rust}
\alias{jackknife_plus_fregre_lm_rust}
\title{Jackknife+ regression using functional linear model}
\usage{
jackknife_plus_fregre_lm_rust(
  data,
  y,
  test_data,
  scalar_train,
  scalar_test,
  ncomp,
  alpha
)
}
\description{
Jackknife+ regression using functional linear model
}
