% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{jackknife_plus_fregre_np_rust}
\alias{jackknife_plus_fregre_np_rust}
\title{Jackknife+ regression using nonparametric kernel regression}
\usage{
jackknife_plus_fregre_np_rust(
  data,
  y,
  test_data,
  argvals,
  scalar_train,
  scalar_test,
  h_func,
  h_scalar,
  alpha
)
}
\description{
Jackknife+ regression using nonparametric kernel regression
}
