% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{knn_gcv}
\alias{knn_gcv}
\title{k-NN with Global Cross-Validation
Finds a single optimal k for all observations}
\usage{
knn_gcv(dist_matrix, response, max_k)
}
\description{
k-NN with Global Cross-Validation
Finds a single optimal k for all observations
}
