% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{knn_lcv}
\alias{knn_lcv}
\title{k-NN with Local Cross-Validation
Finds an optimal k for each observation}
\usage{
knn_lcv(dist_matrix, response, max_k)
}
\description{
k-NN with Local Cross-Validation
Finds an optimal k for each observation
}
