% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{knn_predict}
\alias{knn_predict}
\title{Kernel prediction with fixed bandwidth for prediction on new data}
\usage{
knn_predict(dist_matrix, response, k, local_k)
}
\description{
Kernel prediction with fixed bandwidth for prediction on new data
}
