% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{landmark_detect}
\alias{landmark_detect}
\title{Detect landmarks in a single curve}
\usage{
landmark_detect(curve, argvals, kind, min_prominence)
}
\description{
Detect landmarks in a single curve
}
