% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{landmark_register_curves}
\alias{landmark_register_curves}
\title{Detect landmarks and register curves}
\usage{
landmark_register_curves(data, argvals, kind, min_prominence, expected_count)
}
\description{
Detect landmarks and register curves
}
