% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{metric.elastic}
\alias{metric.elastic}
\title{Elastic Distance (Metric Dispatcher Alias)}
\usage{
metric.elastic(fdataobj, fdataref = NULL, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{fdataref}{Optional reference 'fdata'. If NULL, computes self-distances.}

\item{...}{Additional arguments (ignored).}
}
\value{
A distance matrix (numeric matrix).
}
\description{
Alias for \code{\link{elastic.distance}} for use with the \code{\link{metric}}
dispatcher.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(200), 20, 10), argvals = seq(0, 1, length.out = 10))
D <- metric.elastic(fd)
}

}
