% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_dtw_cross_1d}
\alias{metric_dtw_cross_1d}
\title{Compute DTW distance matrix for cross-distances (n1 x n2)}
\usage{
metric_dtw_cross_1d(fdata1, fdata2, p, w)
}
\description{
Compute DTW distance matrix for cross-distances (n1 x n2)
}
