% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_dtw_self_1d}
\alias{metric_dtw_self_1d}
\title{Compute DTW distance matrix for self-distances (symmetric)}
\usage{
metric_dtw_self_1d(fdata, p, w)
}
\description{
Compute DTW distance matrix for self-distances (symmetric)
}
