% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_hausdorff_cross_1d}
\alias{metric_hausdorff_cross_1d}
\title{Compute Hausdorff distance matrix for cross-distances (n1 x n2)}
\usage{
metric_hausdorff_cross_1d(fdata1, fdata2, argvals)
}
\description{
The Hausdorff distance treats curves as sets of points (t, f(t)) in 2D space.
}
