% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_hausdorff_cross_2d}
\alias{metric_hausdorff_cross_2d}
\title{Compute Hausdorff cross-distances for 2D functional data}
\usage{
metric_hausdorff_cross_2d(fdata1, fdata2, argvals_s, argvals_t)
}
\description{
Compute Hausdorff cross-distances for 2D functional data
}
