% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_kl_cross_1d}
\alias{metric_kl_cross_1d}
\title{Compute symmetric KL divergence matrix for cross-distances (1D)}
\usage{
metric_kl_cross_1d(fdata1, fdata2, argvals, eps, normalize)
}
\description{
Compute symmetric KL divergence matrix for cross-distances (1D)
}
