% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_kl_self_1d}
\alias{metric_kl_self_1d}
\title{Compute symmetric KL divergence matrix for self-distances (1D)
Curves are first normalized to be valid probability distributions}
\usage{
metric_kl_self_1d(fdata, argvals, eps, normalize)
}
\description{
Compute symmetric KL divergence matrix for self-distances (1D)
Curves are first normalized to be valid probability distributions
}
