% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_lp_1d}
\alias{metric_lp_1d}
\title{Compute Lp distance matrix between two sets of functional data}
\usage{
metric_lp_1d(fdata1, fdata2, argvals, p, w)
}
\description{
Compute Lp distance matrix between two sets of functional data
}
