% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_lp_2d}
\alias{metric_lp_2d}
\title{Compute Lp distance between two 2D functional data objects (surfaces)}
\usage{
metric_lp_2d(fdata1, fdata2, argvals_s, argvals_t, p, w)
}
\description{
For 2D functional data, the Lp distance is computed as:
d(f, g) = (∫∫ |f(s,t) - g(s,t)|^p ds dt)^(1/p)
}
\details{
Data is stored as flattened matrices: n x (m1 * m2) where m1 = len(argvals_s), m2 = len(argvals_t)
Data is stored in row-major order for the surface: f(s_i, t_j) is at index i * m2 + j
}
