% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_lp_self_1d}
\alias{metric_lp_self_1d}
\title{Compute Lp distance matrix for self-distances (symmetric)}
\usage{
metric_lp_self_1d(fdata, argvals, p, w)
}
\description{
Compute Lp distance matrix for self-distances (symmetric)
}
