% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_lp_self_2d}
\alias{metric_lp_self_2d}
\title{Compute Lp self-distance matrix for 2D functional data (symmetric)}
\usage{
metric_lp_self_2d(fdata, argvals_s, argvals_t, p, w)
}
\description{
Compute Lp self-distance matrix for 2D functional data (symmetric)
}
