% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_soft_dtw_barycenter}
\alias{metric_soft_dtw_barycenter}
\title{Soft-DTW barycenter computation}
\usage{
metric_soft_dtw_barycenter(fdata, gamma, max_iter, tol)
}
\description{
Soft-DTW barycenter computation
}
