% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_soft_dtw_div_cross_1d}
\alias{metric_soft_dtw_div_cross_1d}
\title{Soft-DTW divergence cross-distance matrix}
\usage{
metric_soft_dtw_div_cross_1d(fdata1, fdata2, gamma)
}
\description{
Soft-DTW divergence cross-distance matrix
}
