% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{metric_soft_dtw_div_self_1d}
\alias{metric_soft_dtw_div_self_1d}
\title{Soft-DTW divergence self-distance matrix}
\usage{
metric_soft_dtw_div_self_1d(fdata, gamma)
}
\description{
Soft-DTW divergence self-distance matrix
}
