% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{model_selection_ncomp_rust}
\alias{model_selection_ncomp_rust}
\title{Model selection for ncomp via AIC/BIC/GCV}
\usage{
model_selection_ncomp_rust(data, y, scalar_covariates, max_ncomp, criterion)
}
\description{
Model selection for ncomp via AIC/BIC/GCV
}
