% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{outliers_lrt}
\alias{outliers_lrt}
\title{LRT-based outlier detection
Returns indices of detected outliers}
\usage{
outliers_lrt(data, argvals, nb, smo, trim, seed, percentile)
}
\description{
LRT-based outlier detection
Returns indices of detected outliers
}
