% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{outliers_thres_lrt}
\alias{outliers_thres_lrt}
\title{Compute bootstrap threshold for LRT outlier detection
Highly parallelized across bootstrap iterations}
\usage{
outliers_thres_lrt(data, argvals, nb, smo, trim, seed, percentile)
}
\description{
Compute bootstrap threshold for LRT outlier detection
Highly parallelized across bootstrap iterations
}
