% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{outliers_thres_lrt_with_dist_rust}
\alias{outliers_thres_lrt_with_dist_rust}
\title{LRT outlier threshold with full bootstrap distribution}
\usage{
outliers_thres_lrt_with_dist_rust(
  data,
  `_argvals`,
  nb,
  smo,
  trim,
  seed,
  percentile
)
}
\description{
LRT outlier threshold with full bootstrap distribution
}
