% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.cluster.gmm}
\alias{plot.cluster.gmm}
\title{Plot Method for cluster.gmm Objects}
\usage{
\method{plot}{cluster.gmm}(x, type = c("bic", "membership"), ...)
}
\arguments{
\item{x}{An object of class 'cluster.gmm'.}

\item{type}{Type of plot: "bic" (default) or "membership".}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plots BIC/ICL model selection curve and cluster memberships.
}
