% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{plot.elastic.align}
\alias{plot.elastic.align}
\title{Plot Elastic Alignment Results}
\usage{
\method{plot}{elastic.align}(x, type = c("aligned", "original", "warps", "both"), ...)
}
\arguments{
\item{x}{An object of class 'elastic.align'.}

\item{type}{Character: "aligned" (default), "original", "warps", or "both".}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object (invisibly).
}
\description{
Plot Elastic Alignment Results
}
