% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_clustering.R
\name{plot.elastic.hclust}
\alias{plot.elastic.hclust}
\title{Plot Elastic Hierarchical Clustering Result}
\usage{
\method{plot}{elastic.hclust}(x, ...)
}
\arguments{
\item{x}{An object of class \code{elastic.hclust}.}

\item{...}{Additional arguments passed to \code{plot.hclust}.}
}
\value{
Invisible NULL. Called for its side effect of producing a plot.
}
\description{
Displays a dendrogram of the hierarchical clustering. The merge
information is converted to an \code{hclust} object for plotting.
}
