% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_clustering.R
\name{plot.elastic.kmeans}
\alias{plot.elastic.kmeans}
\title{Plot Elastic K-Means Result}
\usage{
\method{plot}{elastic.kmeans}(x, ...)
}
\arguments{
\item{x}{An object of class \code{elastic.kmeans}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Displays cluster mean curves with member curves in the background,
one facet per cluster.
}
