% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{plot.fanova}
\alias{plot.fanova}
\title{Plot method for fanova objects}
\usage{
\method{plot}{fanova}(x, ...)
}
\arguments{
\item{x}{A fanova object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plot method for fanova objects
}
