% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{plot.fclassif}
\alias{plot.fclassif}
\title{Plot method for fclassif objects}
\usage{
\method{plot}{fclassif}(x, ...)
}
\arguments{
\item{x}{A fclassif object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plots the confusion matrix as a heatmap.
}
