% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmm.R
\name{plot.fmm}
\alias{plot.fmm}
\title{Plot method for fmm objects}
\usage{
\method{plot}{fmm}(x, ...)
}
\arguments{
\item{x}{A fmm object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plot method for fmm objects
}
