% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{plot.fosr}
\alias{plot.fosr}
\title{Plot method for fosr objects}
\usage{
\method{plot}{fosr}(x, ...)
}
\arguments{
\item{x}{A fosr object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plot method for fosr objects
}
