% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregFdata.R
\name{plot.irregFdata}
\alias{plot.irregFdata}
\title{Plot method for irregFdata objects}
\usage{
\method{plot}{irregFdata}(x, ..., alpha = 0.7)
}
\arguments{
\item{x}{An irregFdata object.}

\item{...}{Additional arguments (ignored).}

\item{alpha}{Transparency for lines.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot method for irregFdata objects
}
