% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{plot.lomb_scargle_result}
\alias{plot.lomb_scargle_result}
\title{Plot method for lomb_scargle_result objects}
\usage{
\method{plot}{lomb_scargle_result}(x, ...)
}
\arguments{
\item{x}{A lomb_scargle_result object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plot method for lomb_scargle_result objects
}
