% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{plot.shape.mean}
\alias{plot.shape.mean}
\title{Plot Shape Mean}
\usage{
\method{plot}{shape.mean}(x, ...)
}
\arguments{
\item{x}{An object of class \code{shape.mean}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Displays the shape mean curve with aligned input curves in the background.
}
