% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{plot.spm.amewma}
\alias{plot.spm.amewma}
\title{Plot AMEWMA monitoring results}
\usage{
\method{plot}{spm.amewma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spm.amewma}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays AMEWMA T-squared statistic and adaptive lambda over time in
a two-panel plot.
}
