% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{plot.spm.chart}
\alias{plot.spm.chart}
\title{Plot an SPM Phase I control chart}
\usage{
\method{plot}{spm.chart}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spm.chart}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays T-squared and SPE statistics for Phase I observations with
upper control limits.
}
