% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{plot.spm.cusum}
\alias{plot.spm.cusum}
\title{Plot CUSUM monitoring results}
\usage{
\method{plot}{spm.cusum}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spm.cusum}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays CUSUM statistic and SPE over time with control limits.
}
