% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{plot.spm.elastic.chart}
\alias{plot.spm.elastic.chart}
\title{Plot Elastic SPM Phase I Chart}
\usage{
\method{plot}{spm.elastic.chart}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spm.elastic.chart}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays amplitude and (optionally) phase control charts for elastic
SPM Phase I.
}
