% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{plot.spm.mewma}
\alias{plot.spm.mewma}
\title{Plot MEWMA monitoring results}
\usage{
\method{plot}{spm.mewma}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spm.mewma}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays MEWMA statistic and SPE over time with control limits.
}
