% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{plot.spm.monitor}
\alias{plot.spm.monitor}
\title{Plot SPM Phase II monitoring results}
\usage{
\method{plot}{spm.monitor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spm.monitor}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays T-squared and SPE monitoring statistics with upper control limits.
Alarm observations are highlighted in red.
}
