% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrvf.R
\name{plot.tsrvf}
\alias{plot.tsrvf}
\title{Plot TSRVF Results}
\usage{
\method{plot}{tsrvf}(x, type = c("tangent", "mean", "gammas"), ...)
}
\arguments{
\item{x}{An object of class 'tsrvf'.}

\item{type}{Character: "tangent" (default), "mean", or "gammas".}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object (invisibly).
}
\description{
Plot TSRVF Results
}
