% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{plot.warp.statistics}
\alias{plot.warp.statistics}
\title{Plot Warp Statistics}
\usage{
\method{plot}{warp.statistics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{warp.statistics}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Displays the mean warping function with a shaded +/- 2 standard deviation
confidence band.
}
