% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{predict.cluster.gmm}
\alias{predict.cluster.gmm}
\title{Predict Cluster Membership for New Functional Data}
\usage{
\method{predict}{cluster.gmm}(object, newdata, new.covariates = NULL, ...)
}
\arguments{
\item{object}{A fitted 'cluster.gmm' object.}

\item{newdata}{An fdata object of new curves.}

\item{new.covariates}{Optional matrix of covariates for new data.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list with \code{cluster} (assignments) and \code{membership} (probabilities).
}
\description{
Assigns new functional observations to clusters using a fitted GMM.
}
