% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_regression.R
\name{predict.elastic.logistic}
\alias{predict.elastic.logistic}
\title{Predict from Elastic Logistic Classification}
\usage{
\method{predict}{elastic.logistic}(object, newdata, type = c("response", "class"), ...)
}
\arguments{
\item{object}{A fitted object of class 'elastic.logistic'.}

\item{newdata}{New functional data (fdata object).}

\item{type}{"response" for probabilities (default) or "class" for classes.}

\item{...}{Additional arguments (ignored).}
}
\value{
Predicted probabilities or class labels.
}
\description{
Predict from Elastic Logistic Classification
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
y <- sample(0:1, 50, replace = TRUE)
fit <- elastic.logistic(fd, y)
predict(fit, fd)
}

}
