% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_regression.R
\name{predict.elastic.regression}
\alias{predict.elastic.regression}
\title{Predict from Elastic Regression}
\usage{
\method{predict}{elastic.regression}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted object of class 'elastic.regression'.}

\item{newdata}{New functional data (fdata object).}

\item{...}{Additional arguments (ignored).}
}
\value{
Predicted response values.
}
\description{
Predict from Elastic Regression
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
y <- rnorm(50)
fit <- elastic.regression(fd, y)
predict(fit, fd)
}

}
