% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{predict.fosr}
\alias{predict.fosr}
\title{Predict from Function-on-Scalar Regression}
\usage{
\method{predict}{fosr}(object, new.predictors = NULL, ...)
}
\arguments{
\item{object}{A fitted 'fosr' object.}

\item{new.predictors}{Matrix of new scalar predictors. If NULL, returns fitted.}

\item{...}{Additional arguments (ignored).}
}
\value{
An fdata object of predicted functional values.
}
\description{
Predict from Function-on-Scalar Regression
}
