% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{predict.fregre.lm}
\alias{predict.fregre.lm}
\title{Predict method for fregre.lm objects}
\usage{
\method{predict}{fregre.lm}(object, newdata = NULL, new.scalar = NULL, ...)
}
\arguments{
\item{object}{A fregre.lm object.}

\item{newdata}{An fdata object (or NULL for fitted values).}

\item{new.scalar}{Optional new scalar covariates matrix.}

\item{...}{Additional arguments (ignored).}
}
\value{
Predicted values.
}
\description{
Predict method for fregre.lm objects
}
