% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{predict.fregre.robust}
\alias{predict.fregre.robust}
\title{Predict from Robust Functional Regression}
\usage{
\method{predict}{fregre.robust}(object, newdata, scalar.covariates = NULL, ...)
}
\arguments{
\item{object}{A fitted 'fregre.robust' model.}

\item{newdata}{An object of class 'fdata' (new curves).}

\item{scalar.covariates}{Optional new scalar covariates.}

\item{...}{Additional arguments (ignored).}
}
\value{
Predicted response values.
}
\description{
Predict from Robust Functional Regression
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
y <- rnorm(50)
fit <- fregre.l1(fd, y, ncomp = 3)
pred <- predict(fit, fd[1:5, ])
}

}
