% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{predict_elastic_logistic_rust}
\alias{predict_elastic_logistic_rust}
\title{Predict from elastic logistic regression}
\usage{
predict_elastic_logistic_rust(
  alpha,
  beta,
  probabilities,
  predicted_classes,
  accuracy,
  loss,
  gammas_data,
  aligned_srsfs_data,
  n_iter,
  new_data,
  argvals
)
}
\description{
Predict from elastic logistic regression
}
