% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{predict_elastic_regression_rust}
\alias{predict_elastic_regression_rust}
\title{Predict from elastic regression}
\usage{
predict_elastic_regression_rust(
  alpha,
  beta,
  fitted_values,
  residuals,
  sse,
  r_squared,
  gammas_data,
  aligned_srsfs_data,
  n_iter,
  new_data,
  argvals
)
}
\description{
Predict from elastic regression
}
