% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{predict_fosr_2d_rust}
\alias{predict_fosr_2d_rust}
\title{Predict from 2D function-on-scalar regression}
\usage{
predict_fosr_2d_rust(
  intercept,
  beta_data,
  new_predictors,
  argvals_s,
  argvals_t,
  lambda_s,
  lambda_t
)
}
\description{
Predict from 2D function-on-scalar regression
}
