% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{predict_fosr_rust}
\alias{predict_fosr_rust}
\title{Predict from function-on-scalar regression}
\usage{
predict_fosr_rust(intercept, beta_data, new_predictors, lambda)
}
\description{
Predict from function-on-scalar regression
}
