% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{predict_fregre_robust_rust}
\alias{predict_fregre_robust_rust}
\title{Predict from robust (L1/Huber) regression}
\usage{
predict_fregre_robust_rust(
  intercept,
  coefficients,
  fpca_mean,
  fpca_rotation_data,
  fpca_rotation_nrow,
  fpca_rotation_ncol,
  ncomp,
  new_data,
  new_scalar
)
}
\description{
Predict from robust (L1/Huber) regression
}
