% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{predict_gmm_rust}
\alias{predict_gmm_rust}
\title{Predict from GMM}
\usage{
predict_gmm_rust(
  new_data,
  argvals,
  new_covariates,
  means_data,
  covariances,
  weights,
  k,
  d,
  nbasis,
  basis_type,
  cov_weight,
  cov_type
)
}
\description{
Predict from GMM
}
