% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{predict_scalar_on_shape_rust}
\alias{predict_scalar_on_shape_rust}
\title{Predict from a scalar-on-shape model}
\usage{
predict_scalar_on_shape_rust(
  beta,
  beta_coefficients,
  gammas,
  shape_scores,
  h_coefficients,
  g_coefficients,
  fitted_values,
  residuals,
  sse,
  r_squared,
  n_iter_outer,
  n_iter_inner,
  index_method,
  index_param,
  new_data,
  argvals
)
}
\description{
Predict from a scalar-on-shape model
}
