% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{print.alignment.diagnostics}
\alias{print.alignment.diagnostics}
\title{Print Alignment Diagnostics}
\usage{
\method{print}{alignment.diagnostics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alignment.diagnostics}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Alignment Diagnostics
}
