% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.cluster.gmm}
\alias{print.cluster.gmm}
\title{Print Method for cluster.gmm Objects}
\usage{
\method{print}{cluster.gmm}(x, ...)
}
\arguments{
\item{x}{An object of class 'cluster.gmm'.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for cluster.gmm Objects
}
