% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{print.cv.fdata}
\alias{print.cv.fdata}
\title{Print Method for cv.fdata}
\usage{
\method{print}{cv.fdata}(x, ...)
}
\arguments{
\item{x}{A \code{cv.fdata} object.}

\item{...}{Ignored.}
}
\description{
Print Method for cv.fdata
}
