% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_clustering.R
\name{print.elastic.hclust}
\alias{print.elastic.hclust}
\title{Print Elastic Hierarchical Clustering Result}
\usage{
\method{print}{elastic.hclust}(x, ...)
}
\arguments{
\item{x}{An object of class \code{elastic.hclust}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Elastic Hierarchical Clustering Result
}
