% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{print.fanova}
\alias{print.fanova}
\title{Print method for fanova objects}
\usage{
\method{print}{fanova}(x, ...)
}
\arguments{
\item{x}{A fanova object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fanova objects
}
