% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{print.fclassif.cv}
\alias{print.fclassif.cv}
\title{Print method for fclassif.cv objects}
\usage{
\method{print}{fclassif.cv}(x, ...)
}
\arguments{
\item{x}{A fclassif.cv object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fclassif.cv objects
}
