% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{print.fclassif}
\alias{print.fclassif}
\title{Print method for fclassif objects}
\usage{
\method{print}{fclassif}(x, ...)
}
\arguments{
\item{x}{A fclassif object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fclassif objects
}
