% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{print.fdata.bootstrap.ci}
\alias{print.fdata.bootstrap.ci}
\title{Print method for bootstrap CI}
\usage{
\method{print}{fdata.bootstrap.ci}(x, ...)
}
\arguments{
\item{x}{A fdata.bootstrap.ci object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for bootstrap CI
}
