% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{print.fdata}
\alias{print.fdata}
\title{Print method for fdata objects}
\usage{
\method{print}{fdata}(x, ...)
}
\arguments{
\item{x}{An object of class 'fdata'.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for fdata objects
}
