% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmm.R
\name{print.fmm}
\alias{print.fmm}
\title{Print method for fmm objects}
\usage{
\method{print}{fmm}(x, ...)
}
\arguments{
\item{x}{A fmm object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fmm objects
}
