% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmm.R
\name{print.fmm.test}
\alias{print.fmm.test}
\title{Print method for fmm.test objects}
\usage{
\method{print}{fmm.test}(x, ...)
}
\arguments{
\item{x}{A fmm.test object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fmm.test objects
}
