% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{print.fosr}
\alias{print.fosr}
\title{Print method for fosr objects}
\usage{
\method{print}{fosr}(x, ...)
}
\arguments{
\item{x}{A fosr object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fosr objects
}
