% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{print.fregre.lm}
\alias{print.fregre.lm}
\title{Print method for fregre.lm objects}
\usage{
\method{print}{fregre.lm}(x, ...)
}
\arguments{
\item{x}{A fregre.lm object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fregre.lm objects
}
