% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{print.fregre.logistic}
\alias{print.fregre.logistic}
\title{Print method for fregre.logistic objects}
\usage{
\method{print}{fregre.logistic}(x, ...)
}
\arguments{
\item{x}{A fregre.logistic object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fregre.logistic objects
}
