% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{print.lambda.cv}
\alias{print.lambda.cv}
\title{Print Lambda CV Result}
\usage{
\method{print}{lambda.cv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lambda.cv}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Lambda CV Result
}
