% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{print.phase.boxplot}
\alias{print.phase.boxplot}
\title{Print Phase Boxplot}
\usage{
\method{print}{phase.boxplot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{phase.boxplot}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Phase Boxplot
}
