% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_on_shape.R
\name{print.scalar.on.shape}
\alias{print.scalar.on.shape}
\title{Print Scalar-on-Shape Model}
\usage{
\method{print}{scalar.on.shape}(x, ...)
}
\arguments{
\item{x}{A \code{scalar.on.shape} object.}

\item{...}{Ignored.}
}
\description{
Print Scalar-on-Shape Model
}
