% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{print.shape.mean}
\alias{print.shape.mean}
\title{Print Shape Mean}
\usage{
\method{print}{shape.mean}(x, ...)
}
\arguments{
\item{x}{An object of class \code{shape.mean}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Shape Mean
}
