% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{print.warp.statistics}
\alias{print.warp.statistics}
\title{Print Warp Statistics}
\usage{
\method{print}{warp.statistics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{warp.statistics}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Warp Statistics
}
