% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pspline_fit_1d}
\alias{pspline_fit_1d}
\title{P-spline fitting: returns coefficients, fitted values, and diagnostics}
\usage{
pspline_fit_1d(data, argvals, nbasis, lambda, order)
}
\description{
P-spline fitting: returns coefficients, fitted values, and diagnostics
}
