% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pspline_fit_2d}
\alias{pspline_fit_2d}
\title{2D P-spline fitting with anisotropic penalties}
\usage{
pspline_fit_2d(
  data,
  argvals_s,
  argvals_t,
  nbasis_s,
  nbasis_t,
  lambda_s,
  lambda_t,
  order
)
}
\description{
2D P-spline fitting with anisotropic penalties
}
