% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{register_shift_1d}
\alias{register_shift_1d}
\title{Shift registration: find optimal horizontal shift for each curve
to align with a target (usually the mean)}
\usage{
register_shift_1d(data, target, argvals, max_shift)
}
\description{
Shift registration: find optimal horizontal shift for each curve
to align with a target (usually the mean)
}
