% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rp_stat}
\alias{rp_stat}
\title{Compute random projection statistics (parallelized over projections)}
\usage{
rp_stat(proj_x_ord, residuals, n_proj)
}
\description{
Compute random projection statistics (parallelized over projections)
}
