% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{s_llr}
\alias{s_llr}
\title{Local Linear Regression smoother matrix
Uses weighted least squares with degree-1 polynomial}
\usage{
s_llr(argvals, h, kernel_type, weights, cv)
}
\description{
Local Linear Regression smoother matrix
Uses weighted least squares with degree-1 polynomial
}
