% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{s_lpr}
\alias{s_lpr}
\title{Local Polynomial Regression smoother matrix
Solves (p+1)×(p+1) weighted least squares system for each point}
\usage{
s_lpr(argvals, h, p, kernel_type, weights, cv)
}
\description{
Local Polynomial Regression smoother matrix
Solves (p+1)×(p+1) weighted least squares system for each point
}
