% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{s_nw}
\alias{s_nw}
\title{Nadaraya-Watson smoother matrix
S_ij = K((t_i - t_j)/h) * w_j / sum_k(K((t_i - t_k)/h) * w_k)}
\usage{
s_nw(argvals, h, kernel_type, weights, cv)
}
\description{
Nadaraya-Watson smoother matrix
S_ij = K((t_i - t_j)/h) * w_j / sum_k(K((t_i - t_k)/h) * w_k)
}
