% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{scalar_on_shape_rust}
\alias{scalar_on_shape_rust}
\title{Fit scalar-on-shape regression model}
\usage{
scalar_on_shape_rust(
  data,
  y,
  argvals,
  nbasis,
  lambda,
  index_method,
  index_param,
  g_degree,
  dp_lambda,
  max_iter_outer,
  max_iter_inner,
  tol
)
}
\description{
Fit scalar-on-shape regression model
}
