% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_analyze_peak_timing}
\alias{seasonal_analyze_peak_timing}
\title{Analyze peak timing variability across cycles (uses Fourier smoothing)}
\usage{
seasonal_analyze_peak_timing(data, argvals, period, smooth_nbasis)
}
\description{
Analyze peak timing variability across cycles (uses Fourier smoothing)
}
