% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_cfd_autoperiod}
\alias{seasonal_cfd_autoperiod}
\title{CFDAutoperiod: Clustered Filtered Detrended Autoperiod
Uses differencing for detrending and clustering for robust period detection}
\usage{
seasonal_cfd_autoperiod(data, argvals, cluster_tolerance, min_cluster_size)
}
\description{
CFDAutoperiod: Clustered Filtered Detrended Autoperiod
Uses differencing for detrending and clustering for robust period detection
}
