% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_classify_seasonality}
\alias{seasonal_classify_seasonality}
\title{Classify seasonality type}
\usage{
seasonal_classify_seasonality(
  data,
  argvals,
  period,
  strength_threshold,
  timing_threshold
)
}
\description{
Classify seasonality type
}
