% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_decompose}
\alias{seasonal_decompose}
\title{Decompose functional data into trend, seasonal, and remainder components}
\usage{
seasonal_decompose(
  data,
  argvals,
  period,
  method,
  trend_method,
  bandwidth,
  n_harmonics
)
}
\description{
Decompose functional data into trend, seasonal, and remainder components
}
