% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_detect_changes}
\alias{seasonal_detect_changes}
\title{Detect seasonality changes (onset/cessation)}
\usage{
seasonal_detect_changes(
  data,
  argvals,
  period,
  threshold,
  window_size,
  min_duration
)
}
\description{
Detect seasonality changes (onset/cessation)
}
