% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_detect_changes_auto}
\alias{seasonal_detect_changes_auto}
\title{Detect seasonality changes with automatic threshold}
\usage{
seasonal_detect_changes_auto(
  data,
  argvals,
  period,
  threshold_method,
  threshold_value,
  window_size,
  min_duration
)
}
\description{
Detect seasonality changes with automatic threshold
}
