% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_detect_multiple_periods}
\alias{seasonal_detect_multiple_periods}
\title{Detect multiple concurrent periodicities using iterative residual subtraction}
\usage{
seasonal_detect_multiple_periods(
  data,
  argvals,
  max_periods,
  min_confidence,
  min_strength
)
}
\description{
Detect multiple concurrent periodicities using iterative residual subtraction
}
