% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_detect_peaks}
\alias{seasonal_detect_peaks}
\title{Detect peaks in functional data using Fourier basis smoothing}
\usage{
seasonal_detect_peaks(
  data,
  argvals,
  min_distance,
  min_prominence,
  smooth_first,
  smooth_nbasis
)
}
\description{
Detect peaks in functional data using Fourier basis smoothing
}
