% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_detrend}
\alias{seasonal_detrend}
\title{Detrend functional data using specified method
Returns trend, detrended data, method used, RSS per curve, and number of parameters}
\usage{
seasonal_detrend(data, argvals, method, degree, bandwidth)
}
\description{
Detrend functional data using specified method
Returns trend, detrended data, method used, RSS per curve, and number of parameters
}
