% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_estimate_period_acf}
\alias{seasonal_estimate_period_acf}
\title{Estimate period using autocorrelation}
\usage{
seasonal_estimate_period_acf(data, argvals, max_lag)
}
\description{
Estimate period using autocorrelation
}
