% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_instantaneous_period}
\alias{seasonal_instantaneous_period}
\title{Estimate instantaneous period using Hilbert transform}
\usage{
seasonal_instantaneous_period(data, argvals)
}
\description{
Estimate instantaneous period using Hilbert transform
}
