% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_lomb_scargle}
\alias{seasonal_lomb_scargle}
\title{Lomb-Scargle periodogram for irregularly sampled data
Computes the power spectrum and significance for period detection}
\usage{
seasonal_lomb_scargle(data, argvals, oversampling, nyquist_factor)
}
\description{
Lomb-Scargle periodogram for irregularly sampled data
Computes the power spectrum and significance for period detection
}
