% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_stl}
\alias{seasonal_stl}
\title{STL (Seasonal and Trend decomposition using LOESS)
Implements Cleveland et al. 1990 algorithm}
\usage{
seasonal_stl(data, period, s_window, t_window, robust)
}
\description{
STL (Seasonal and Trend decomposition using LOESS)
Implements Cleveland et al. 1990 algorithm
}
