% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_strength_spectral}
\alias{seasonal_strength_spectral}
\title{Measure seasonal strength using spectral method}
\usage{
seasonal_strength_spectral(data, argvals, period)
}
\description{
Measure seasonal strength using spectral method
}
