% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_strength_variance}
\alias{seasonal_strength_variance}
\title{Measure seasonal strength using variance decomposition}
\usage{
seasonal_strength_variance(data, argvals, period, n_harmonics)
}
\description{
Measure seasonal strength using variance decomposition
}
