% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_strength_wavelet}
\alias{seasonal_strength_wavelet}
\title{Measure seasonal strength using wavelet (Morlet) method}
\usage{
seasonal_strength_wavelet(data, argvals, period)
}
\arguments{
\item{data}{Matrix of functional data (n x m)}

\item{argvals}{Vector of evaluation points (length m)}

\item{period}{Seasonal period in argvals units}
}
\value{
Seasonal strength as ratio of wavelet power to total variance (0 to 1)
}
\description{
Uses Continuous Wavelet Transform with Morlet wavelet to measure
power at the specified seasonal period.
}
