% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{seasonal_strength_windowed}
\alias{seasonal_strength_windowed}
\title{Time-varying seasonal strength using sliding windows}
\usage{
seasonal_strength_windowed(data, argvals, period, window_size, method)
}
\description{
Time-varying seasonal strength using sliding windows
}
