% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_basis_cross_1d}
\alias{semimetric_basis_cross_1d}
\title{Basis coefficient semimetric (cross-distances)}
\usage{
semimetric_basis_cross_1d(data1, data2, argvals, nbasis, basis_type)
}
\description{
Basis coefficient semimetric (cross-distances)
}
