% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_basis_self_1d}
\alias{semimetric_basis_self_1d}
\title{Basis coefficient semimetric (self-distances)}
\usage{
semimetric_basis_self_1d(data, argvals, nbasis, basis_type)
}
\description{
Basis coefficient semimetric (self-distances)
}
