% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_deriv_cross_1d}
\alias{semimetric_deriv_cross_1d}
\title{Derivative-based semimetric (cross-distances)}
\usage{
semimetric_deriv_cross_1d(data1, data2, argvals, nderiv, weights)
}
\description{
Derivative-based semimetric (cross-distances)
}
