% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_deriv_self_1d}
\alias{semimetric_deriv_self_1d}
\title{Derivative-based semimetric (self-distances)}
\usage{
semimetric_deriv_self_1d(data, argvals, nderiv, weights)
}
\description{
Derivative-based semimetric (self-distances)
}
