% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_fourier_cross_1d}
\alias{semimetric_fourier_cross_1d}
\title{Compute semimetric based on Fourier coefficients for cross-distances}
\usage{
semimetric_fourier_cross_1d(fdata1, fdata2, nfreq)
}
\description{
Compute semimetric based on Fourier coefficients for cross-distances
}
