% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_fourier_self_1d}
\alias{semimetric_fourier_self_1d}
\title{Compute semimetric based on Fourier coefficients for self-distances (symmetric)
Uses FFT to compute Fourier coefficients and then L2 distance on coefficients}
\usage{
semimetric_fourier_self_1d(fdata, nfreq)
}
\description{
Compute semimetric based on Fourier coefficients for self-distances (symmetric)
Uses FFT to compute Fourier coefficients and then L2 distance on coefficients
}
