% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_hshift_cross_1d}
\alias{semimetric_hshift_cross_1d}
\title{Compute semimetric based on horizontal shift for cross-distances}
\usage{
semimetric_hshift_cross_1d(fdata1, fdata2, argvals, max_shift)
}
\description{
Compute semimetric based on horizontal shift for cross-distances
}
