% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_hshift_self_1d}
\alias{semimetric_hshift_self_1d}
\title{Compute semimetric based on horizontal shift for self-distances (symmetric)
This finds the minimum L2 distance after optimally shifting one curve horizontally}
\usage{
semimetric_hshift_self_1d(fdata, argvals, max_shift)
}
\description{
Compute semimetric based on horizontal shift for self-distances (symmetric)
This finds the minimum L2 distance after optimally shifting one curve horizontally
}
