% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_pca_cross_1d}
\alias{semimetric_pca_cross_1d}
\title{PCA-based semimetric (cross-distances)}
\usage{
semimetric_pca_cross_1d(data1, data2, ncomp)
}
\description{
PCA-based semimetric (cross-distances)
}
