% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{semimetric_pca_self_1d}
\alias{semimetric_pca_self_1d}
\title{PCA-based semimetric (self-distances)}
\usage{
semimetric_pca_self_1d(data, ncomp)
}
\description{
PCA-based semimetric (self-distances)
}
