% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sim_kl_1d}
\alias{sim_kl_1d}
\title{Simulate functional data via Karhunen-Loève expansion}
\usage{
sim_kl_1d(n, phi, lambda, seed)
}
\description{
Simulate functional data via Karhunen-Loève expansion
}
