% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{smooth_basis_gcv_rust}
\alias{smooth_basis_gcv_rust}
\title{Smooth functional data with GCV-selected penalty}
\usage{
smooth_basis_gcv_rust(
  data,
  argvals,
  basis_type,
  nbasis,
  lfd_order,
  log_lambda_min,
  log_lambda_max,
  n_grid,
  period
)
}
\description{
Smooth functional data with GCV-selected penalty
}
