% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{softdtw.barycenter}
\alias{softdtw.barycenter}
\title{Soft-DTW Barycenter}
\usage{
softdtw.barycenter(fdataobj, gamma = 1, max.iter = 100, tol = 1e-05)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{gamma}{Smoothing parameter (> 0). Default is 1.0.}

\item{max.iter}{Maximum number of iterations. Default is 100.}

\item{tol}{Convergence tolerance. Default is 1e-5.}
}
\value{
An object of class 'fdata' containing the barycenter curve,
with attributes \code{n.iter} and \code{converged}.
}
\description{
Compute the barycenter (weighted average) of functional data using the
Soft-DTW distance. The barycenter minimizes the sum of Soft-DTW distances
to all input curves.
}
\examples{
fd <- fdata(matrix(rnorm(200), 20, 10), argvals = seq(0, 1, length.out = 10))
bc <- softdtw.barycenter(fd, gamma = 1.0)
}
\references{
Cuturi, M. and Blondel, M. (2017). Soft-DTW: a Differentiable Loss Function
for Time-Series. \emph{Proceedings of the 34th International Conference on
Machine Learning (ICML)}.
}
