% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{spm.elastic.monitor}
\alias{spm.elastic.monitor}
\title{Elastic SPM Phase II Monitoring}
\usage{
spm.elastic.monitor(chart, newdata)
}
\arguments{
\item{chart}{An object of class \code{spm.elastic.chart} from
\code{\link{spm.elastic.phase1}}.}

\item{newdata}{An object of class \code{fdata} with new observations.}
}
\value{
A list with components:
\describe{
\item{amplitude}{List with t2, spe, t2.alarm, spe.alarm, scores}
\item{phase}{List with t2, spe, t2.alarm, spe.alarm, scores (if
\code{monitor.phase = TRUE}), or NULL}
\item{aligned.data}{Matrix of aligned curves}
\item{warping.functions}{Matrix of estimated warping functions}
}
}
\description{
Monitors new functional data against an elastic SPM chart. New curves are
aligned to the Karcher mean, then amplitude (and optionally phase) are
monitored separately.
}
\examples{
\donttest{
set.seed(1)
n <- 40; m <- 30
argvals <- seq(0, 1, length.out = m)
X <- matrix(rnorm(n * m), n, m)
fd <- fdata(X, argvals = argvals)
chart <- spm.elastic.phase1(fd, ncomp = 3, warp.ncomp = 2)

X_new <- matrix(rnorm(10 * m) + 1, 10, m)
fd_new <- fdata(X_new, argvals = argvals)
mon <- spm.elastic.monitor(chart, fd_new)
mon$amplitude$t2.alarm
}

}
\seealso{
\code{\link{spm.elastic.phase1}} for building the chart
}
