% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_amewma_monitor_rust}
\alias{spm_amewma_monitor_rust}
\title{Adaptive EWMA (AMEWMA) monitoring on sequential functional data.}
\usage{
spm_amewma_monitor_rust(
  sequential_data,
  argvals,
  rotation,
  mean,
  singular_values,
  centered,
  eigenvalues,
  t2_ucl,
  t2_alpha,
  t2_description,
  spe_ucl,
  spe_alpha,
  spe_description,
  chart_ncomp,
  config_alpha,
  config_tuning_fraction,
  config_seed,
  lambda_min,
  lambda_max,
  lambda_init,
  eta,
  amewma_ncomp,
  amewma_alpha
)
}
\description{
Adaptive EWMA (AMEWMA) monitoring on sequential functional data.
}
