% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_arl0_spe_rust}
\alias{spm_arl0_spe_rust}
\title{In-control ARL for SPE chart.}
\usage{
spm_arl0_spe_rust(spe_df, spe_scale, ucl, n_sim, max_rl, seed)
}
\description{
In-control ARL for SPE chart.
}
