% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_arl0_t2_rust}
\alias{spm_arl0_t2_rust}
\title{In-control ARL for T-squared chart.}
\usage{
spm_arl0_t2_rust(eigenvalues, ucl, n_sim, max_rl, seed)
}
\description{
In-control ARL for T-squared chart.
}
