% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_arl1_t2_rust}
\alias{spm_arl1_t2_rust}
\title{Out-of-control ARL for T-squared chart.}
\usage{
spm_arl1_t2_rust(eigenvalues, ucl, shift, n_sim, max_rl, seed)
}
\description{
Out-of-control ARL for T-squared chart.
}
