% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_cusum_monitor_rust}
\alias{spm_cusum_monitor_rust}
\title{CUSUM monitoring on sequential functional data.}
\usage{
spm_cusum_monitor_rust(
  sequential_data,
  argvals,
  rotation,
  mean,
  singular_values,
  centered,
  eigenvalues,
  t2_ucl,
  t2_alpha,
  t2_description,
  spe_ucl,
  spe_alpha,
  spe_description,
  chart_ncomp,
  config_alpha,
  config_tuning_fraction,
  config_seed,
  k,
  h,
  cusum_ncomp,
  cusum_alpha,
  multivariate
)
}
\description{
CUSUM monitoring on sequential functional data.
}
