% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_evaluate_rules_rust}
\alias{spm_evaluate_rules_rust}
\title{Evaluate control chart rules (Western Electric, Nelson, or custom).}
\usage{
spm_evaluate_rules_rust(values, center, sigma, rule_set)
}
\description{
Evaluate control chart rules (Western Electric, Nelson, or custom).
}
