% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_ewma_rust}
\alias{spm_ewma_rust}
\title{EWMA-based SPM monitoring.}
\usage{
spm_ewma_rust(
  sequential_data,
  argvals,
  rotation,
  mean,
  singular_values,
  centered,
  eigenvalues,
  t2_ucl,
  t2_alpha,
  t2_description,
  spe_ucl,
  spe_alpha,
  spe_description,
  chart_ncomp,
  config_alpha,
  config_tuning_fraction,
  config_seed,
  lambda,
  ewma_ncomp,
  ewma_alpha
)
}
\description{
EWMA-based SPM monitoring.
}
