% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_mewma_monitor_rust}
\alias{spm_mewma_monitor_rust}
\title{MEWMA monitoring on sequential functional data.}
\usage{
spm_mewma_monitor_rust(
  sequential_data,
  argvals,
  rotation,
  mean,
  singular_values,
  centered,
  eigenvalues,
  t2_ucl,
  t2_alpha,
  t2_description,
  spe_ucl,
  spe_alpha,
  spe_description,
  chart_ncomp,
  config_alpha,
  config_tuning_fraction,
  config_seed,
  lambda,
  mewma_ncomp,
  mewma_alpha,
  asymptotic
)
}
\description{
MEWMA monitoring on sequential functional data.
}
