% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_monitor_partial_batch_rust}
\alias{spm_monitor_partial_batch_rust}
\title{Monitor a batch of partially-observed curves.}
\usage{
spm_monitor_partial_batch_rust(
  partial_data_list,
  n_observed_vec,
  argvals,
  rotation,
  mean,
  singular_values,
  centered,
  eigenvalues,
  t2_ucl,
  t2_alpha,
  t2_description,
  spe_ucl,
  spe_alpha,
  spe_description,
  chart_ncomp,
  config_alpha,
  config_tuning_fraction,
  config_seed,
  partial_ncomp,
  partial_alpha,
  completion_method
)
}
\description{
Monitor a batch of partially-observed curves.
}
