% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_monitor_partial_rust}
\alias{spm_monitor_partial_rust}
\title{Monitor a single partially-observed curve.}
\usage{
spm_monitor_partial_rust(
  partial_values,
  argvals,
  n_observed,
  rotation,
  mean,
  singular_values,
  centered,
  eigenvalues,
  t2_ucl,
  t2_alpha,
  t2_description,
  spe_ucl,
  spe_alpha,
  spe_description,
  chart_ncomp,
  config_alpha,
  config_tuning_fraction,
  config_seed,
  partial_ncomp,
  partial_alpha,
  completion_method
)
}
\description{
Monitor a single partially-observed curve.
}
