% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_monitor_rust}
\alias{spm_monitor_rust}
\title{SPM Phase II: Monitor new data against an established chart.}
\usage{
spm_monitor_rust(
  new_data,
  argvals,
  rotation,
  mean,
  singular_values,
  centered,
  eigenvalues,
  t2_ucl,
  t2_alpha,
  t2_description,
  spe_ucl,
  spe_alpha,
  spe_description,
  ncomp,
  config_alpha,
  config_tuning_fraction,
  config_seed
)
}
\description{
SPM Phase II: Monitor new data against an established chart.
}
