% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_phase1_iterative_rust}
\alias{spm_phase1_iterative_rust}
\title{Iterative Phase I chart construction.}
\usage{
spm_phase1_iterative_rust(
  data,
  argvals,
  ncomp,
  alpha,
  tuning_fraction,
  seed,
  max_iterations,
  max_removal_fraction
)
}
\description{
Iterative Phase I chart construction.
}
