% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_phase1_rust}
\alias{spm_phase1_rust}
\title{SPM Phase I: Build a univariate control chart from in-control functional data.}
\usage{
spm_phase1_rust(data, argvals, ncomp, alpha, tuning_fraction, seed)
}
\description{
SPM Phase I: Build a univariate control chart from in-control functional data.
}
