% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_profile_monitor_rust}
\alias{spm_profile_monitor_rust}
\title{Profile monitoring Phase II.}
\usage{
spm_profile_monitor_rust(
  new_y,
  new_predictors,
  argvals,
  beta_rotation,
  beta_mean,
  beta_singular_values,
  beta_centered,
  ref_fosr_intercept,
  ref_fosr_beta,
  ref_fosr_lambda,
  eigenvalues,
  t2_ucl,
  t2_alpha,
  t2_description,
  lag1_autocorrelation,
  effective_n_windows,
  fosr_lambda,
  ncomp,
  alpha,
  window_size,
  step_size
)
}
\description{
Profile monitoring Phase II.
}
