% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_profile_phase1_rust}
\alias{spm_profile_phase1_rust}
\title{Profile monitoring Phase I.}
\usage{
spm_profile_phase1_rust(
  y_curves,
  predictors,
  argvals,
  fosr_lambda,
  ncomp,
  alpha,
  window_size,
  step_size
)
}
\description{
Profile monitoring Phase I.
}
