% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_select_ncomp_rust}
\alias{spm_select_ncomp_rust}
\title{Select the number of principal components.}
\usage{
spm_select_ncomp_rust(eigenvalues, method, threshold)
}
\description{
Select the number of principal components.
}
