% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_spe_contrib_rust}
\alias{spm_spe_contrib_rust}
\title{SPE contribution diagnostics.}
\usage{
spm_spe_contrib_rust(std_list, recon_list, argvals_list)
}
\description{
SPE contribution diagnostics.
}
