% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_spe_limit_robust_rust}
\alias{spm_spe_limit_robust_rust}
\title{Robust SPE control limit.}
\usage{
spm_spe_limit_robust_rust(spe_values, alpha, method, n_bootstrap, seed)
}
\description{
Robust SPE control limit.
}
