% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_t2_contrib_rust}
\alias{spm_t2_contrib_rust}
\title{T-squared contribution diagnostics.}
\usage{
spm_t2_contrib_rust(scores, eigenvalues, grid_sizes)
}
\description{
T-squared contribution diagnostics.
}
