% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_t2_limit_robust_rust}
\alias{spm_t2_limit_robust_rust}
\title{Robust T-squared control limit.}
\usage{
spm_t2_limit_robust_rust(t2_values, ncomp, alpha, method, n_bootstrap, seed)
}
\description{
Robust T-squared control limit.
}
