% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{spm_t2_pc_contrib_rust}
\alias{spm_t2_pc_contrib_rust}
\title{Per-PC T-squared contributions for a single observation or batch.}
\usage{
spm_t2_pc_contrib_rust(scores, eigenvalues)
}
\description{
Per-PC T-squared contributions for a single observation or batch.
}
