% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{srsf.transform}
\alias{srsf.transform}
\title{Elastic Alignment for Functional Data}
\usage{
srsf.transform(fdataobj)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}
}
\value{
An object of class 'fdata' containing the SRSF-transformed curves.
}
\description{
Functions for elastic curve alignment via SRSF transforms, Fisher-Rao distance,
and Karcher mean computation.
SRSF Transform
}
\details{
Compute the Square-Root Slope Function (SRSF) transformation of functional data:
\eqn{q(t) = \text{sign}(f'(t)) \sqrt{|f'(t)|}}.
}
\examples{
fd <- fdata(matrix(rnorm(200), 20, 10), argvals = seq(0, 1, length.out = 10))
q <- srsf.transform(fd)
}
