% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{streaming_depth_batch}
\alias{streaming_depth_batch}
\title{Streaming depth: batch self-depth computation}
\usage{
streaming_depth_batch(data, method)
}
\description{
Streaming depth: batch self-depth computation
}
