% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{streaming_depth_one}
\alias{streaming_depth_one}
\title{Streaming depth: single curve against reference}
\usage{
streaming_depth_one(ref_data, curve, method)
}
\description{
Streaming depth: single curve against reference
}
