% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{streaming_depth_vs_ref}
\alias{streaming_depth_vs_ref}
\title{Streaming depth: new data against reference}
\usage{
streaming_depth_vs_ref(ref_data, new_data, method)
}
\description{
Streaming depth: new data against reference
}
