% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{summary.basis.auto}
\alias{summary.basis.auto}
\title{Summary method for basis.auto objects}
\usage{
\method{summary}{basis.auto}(object, ...)
}
\arguments{
\item{object}{A basis.auto object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for basis.auto objects
}
