% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{tolerance_conformal}
\alias{tolerance_conformal}
\title{Conformal prediction band}
\usage{
tolerance_conformal(data, cal_fraction, coverage, score_type, seed)
}
\description{
Conformal prediction band
}
