% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{tolerance_elastic}
\alias{tolerance_elastic}
\title{Elastic tolerance band (alignment + FPCA)}
\usage{
tolerance_elastic(
  data,
  argvals,
  ncomp,
  nb,
  coverage,
  band_type,
  max_iter,
  seed
)
}
\description{
Elastic tolerance band (alignment + FPCA)
}
