% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{tolerance_elastic_config_rust}
\alias{tolerance_elastic_config_rust}
\title{Elastic tolerance band with full config (amplitude + phase)}
\usage{
tolerance_elastic_config_rust(
  data,
  argvals,
  ncomp_amplitude,
  ncomp_phase,
  nb,
  coverage,
  band_type,
  max_iter,
  tol,
  seed
)
}
\description{
Elastic tolerance band with full config (amplitude + phase)
}
