% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{tolerance_exponential}
\alias{tolerance_exponential}
\title{Exponential family tolerance band}
\usage{
tolerance_exponential(data, family, ncomp, nb, coverage, seed)
}
\description{
Exponential family tolerance band
}
