% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{tolerance_phase_rust}
\alias{tolerance_phase_rust}
\title{Phase tolerance band (warping variation)}
\usage{
tolerance_phase_rust(
  data,
  argvals,
  ncomp,
  nb,
  coverage,
  band_type,
  max_iter,
  seed
)
}
\description{
Phase tolerance band (warping variation)
}
