% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{tolerance_scb_degras}
\alias{tolerance_scb_degras}
\title{SCB mean confidence band (Degras method)}
\usage{
tolerance_scb_degras(data, argvals, bandwidth, nb, confidence, multiplier)
}
\description{
SCB mean confidence band (Degras method)
}
