% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrvf.R
\name{tsrvf.from.alignment}
\alias{tsrvf.from.alignment}
\title{TSRVF from Pre-computed Alignment}
\usage{
tsrvf.from.alignment(karcher)
}
\arguments{
\item{karcher}{An object of class 'karcher.mean'.}
}
\value{
An object of class 'tsrvf'.
}
\description{
Compute the TSRVF representation from a pre-computed Karcher mean,
avoiding the expensive re-computation of the alignment.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(200), 20, 10), argvals = seq(0, 1, length.out = 10))
km <- karcher.mean(fd, max.iter = 5)
tv <- tsrvf.from.alignment(km)
}
}
