% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrvf.R
\name{tsrvf.inverse}
\alias{tsrvf.inverse}
\title{Inverse TSRVF Transform}
\usage{
tsrvf.inverse(tsrvf.obj)
}
\arguments{
\item{tsrvf.obj}{An object of class 'tsrvf'.}
}
\value{
An fdata object of reconstructed aligned curves.
}
\description{
Reconstruct aligned curves from their TSRVF tangent vector representation.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(200), 20, 10), argvals = seq(0, 1, length.out = 10))
tv <- tsrvf.transform(fd, max.iter = 5)
recon <- tsrvf.inverse(tv)
}
}
