% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{warp.inverse}
\alias{warp.inverse}
\title{Invert a Warping Function}
\usage{
warp.inverse(gamma, argvals)
}
\arguments{
\item{gamma}{Numeric vector of warping function values.}

\item{argvals}{Numeric vector of evaluation points.}
}
\value{
Numeric vector of the inverse warping function.
}
\description{
Compute the functional inverse of a warping function \eqn{\gamma},
i.e., find \eqn{\gamma^{-1}} such that
\eqn{\gamma(\gamma^{-1}(t)) = t}.
}
\examples{
t <- seq(0, 1, length.out = 50)
gamma <- t^2
gamma_inv <- warp.inverse(gamma, t)
}
\seealso{
\code{\link{warp.inverse.error}}, \code{\link{warp.compose}}
}
