% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{basis2fdata_1d}
\alias{basis2fdata_1d}
\title{Reconstruct functional data from basis coefficients
Returns data matrix \link{n x m}}
\usage{
basis2fdata_1d(coefs, argvals, nbasis, basis_type)
}
\description{
Reconstruct functional data from basis coefficients
Returns data matrix \link{n x m}
}
