% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R, R/extendr-wrappers.R
\name{basis2fdata_2d}
\alias{basis2fdata_2d}
\title{Reconstruct 2D Functional Data from Tensor Product Basis Coefficients}
\usage{
basis2fdata_2d(coefs, argvals_s, argvals_t, nbasis_s, nbasis_t, basis_type)

basis2fdata_2d(coefs, argvals_s, argvals_t, nbasis_s, nbasis_t, basis_type)
}
\arguments{
\item{coefs}{Coefficient matrix \code{[n x (nbasis.s * nbasis.t)]}.}

\item{argvals}{List with two numeric vectors for s and t coordinates.}

\item{nbasis.s}{Number of basis functions in s direction.}

\item{nbasis.t}{Number of basis functions in t direction.}

\item{type}{Basis type: "bspline" (default) or "fourier".}
}
\value{
A 2D fdata object.
}
\description{
Reconstructs 2D surfaces from tensor product basis coefficients.
}
