% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bootstrap_ci_functional_logistic_rust}
\alias{bootstrap_ci_functional_logistic_rust}
\title{Bootstrap confidence intervals for β(t) in functional logistic model}
\usage{
bootstrap_ci_functional_logistic_rust(
  data,
  y,
  scalar_covariates,
  ncomp,
  n_boot,
  alpha,
  seed,
  max_iter,
  tol
)
}
\description{
Bootstrap confidence intervals for β(t) in functional logistic model
}
