% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{elastic.align}
\alias{elastic.align}
\title{Elastic Curve Alignment}
\usage{
elastic.align(fdataobj, target = NULL)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{target}{Optional target curve (numeric vector). If NULL, uses the
cross-sectional mean as the alignment target.}
}
\value{
An object of class 'elastic.align' with components:
\describe{
\item{aligned}{fdata of aligned curves}
\item{gammas}{fdata of warping functions}
\item{distances}{numeric vector of elastic distances}
\item{target}{the target curve used for alignment}
\item{fdataobj}{the original fdata input}
}
}
\description{
Align a set of functional curves using the elastic (Fisher-Rao) framework.
When no target is specified, the Karcher mean is used as the alignment target.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(200), 20, 10), argvals = seq(0, 1, length.out = 10))
res <- elastic.align(fd)
}
}
