% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_amp_changepoint_rust}
\alias{elastic_amp_changepoint_rust}
\title{Amplitude changepoint detection}
\usage{
elastic_amp_changepoint_rust(data, argvals, lambda, max_iter, n_mc, seed)
}
\description{
Amplitude changepoint detection
}
