% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_joint_fpca_rust}
\alias{elastic_joint_fpca_rust}
\title{Joint FPCA (amplitude + phase)}
\usage{
elastic_joint_fpca_rust(
  aligned_data,
  gammas_data,
  mean_curve,
  mean_srsf,
  aligned_srsfs,
  argvals,
  ncomp
)
}
\description{
Joint FPCA (amplitude + phase)
}
