% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_logistic_rust}
\alias{elastic_logistic_rust}
\title{Elastic logistic regression}
\usage{
elastic_logistic_rust(data, y, argvals, ncomp_beta, lambda, max_iter, tol)
}
\description{
Elastic logistic regression
}
