% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_regression_rust}
\alias{elastic_regression_rust}
\title{Elastic regression (function-on-scalar with alignment)}
\usage{
elastic_regression_rust(data, y, argvals, ncomp_beta, lambda, max_iter, tol)
}
\description{
Elastic regression (function-on-scalar with alignment)
}
