% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{elastic_vert_fpca_rust}
\alias{elastic_vert_fpca_rust}
\title{Vertical (amplitude) FPCA after elastic alignment}
\usage{
elastic_vert_fpca_rust(
  aligned_data,
  gammas,
  mean_data,
  mean_srsf,
  aligned_srsfs,
  argvals,
  ncomp
)
}
\description{
Vertical (amplitude) FPCA after elastic alignment
}
