% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R, R/extendr-wrappers.R
\name{fdata2basis_2d}
\alias{fdata2basis_2d}
\title{Convert 2D Functional Data to Tensor Product Basis Coefficients}
\usage{
fdata2basis_2d(data, argvals_s, argvals_t, nbasis_s, nbasis_t, basis_type)

fdata2basis_2d(data, argvals_s, argvals_t, nbasis_s, nbasis_t, basis_type)
}
\arguments{
\item{fdataobj}{A 2D fdata object (surfaces).}

\item{nbasis.s}{Number of basis functions in the s (first) direction.}

\item{nbasis.t}{Number of basis functions in the t (second) direction.}

\item{type}{Basis type: "bspline" (default) or "fourier".}
}
\value{
A matrix of coefficients \code{[n x (nbasis.s * nbasis.t)]}.
}
\description{
Projects 2D functional data (surfaces) onto a tensor product basis,
which is the Kronecker product of two 1D bases.
}
\details{
The tensor product basis is defined as:
\deqn{B_{2d}(s, t) = B_s(s) \otimes B_t(t)}
}
\examples{
# Create 2D surface data
s <- seq(0, 1, length.out = 20)
t <- seq(0, 1, length.out = 20)
surface <- outer(sin(2*pi*s), cos(2*pi*t))
fd2d <- fdata(array(surface, dim = c(1, 20, 20)))

# Project to tensor product basis
coefs <- fdata2basis_2d(fd2d, nbasis.s = 7, nbasis.t = 7, type = "fourier")
}
