% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmm.R
\name{fmm.predict}
\alias{fmm.predict}
\title{Predict from Functional Mixed Model}
\usage{
fmm.predict(object, new.covariates = NULL, ...)
}
\arguments{
\item{object}{A fitted 'fmm' object.}

\item{new.covariates}{Matrix of new covariate values. If NULL, returns
the overall mean function.}

\item{...}{Additional arguments (ignored).}
}
\value{
An fdata object of predicted functional values.
}
\description{
Predict population-level (fixed effects only) functional responses
for new covariate values.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
subject <- rep(1:10, each = 5)
x <- cbind(rnorm(50))
fit <- fmm(fd, subject.ids = subject, covariates = x)
pred <- fmm.predict(fit, new.covariates = cbind(c(0, 1)))
}

}
