% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{fosr.fpc}
\alias{fosr.fpc}
\title{FPC-based Function-on-Scalar Regression}
\usage{
fosr.fpc(fdataobj, predictors, ncomp = 3)
}
\arguments{
\item{fdataobj}{An object of class 'fdata' (functional response).}

\item{predictors}{A matrix of scalar predictors (n x p).}

\item{ncomp}{Number of FPC components (default 3).}
}
\value{
An object of class 'fosr'.
}
\description{
Fits a function-on-scalar regression using functional principal components.
}
