% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr.R
\name{fosr}
\alias{fosr}
\title{Function-on-Scalar Regression}
\usage{
fosr(fdataobj, predictors, lambda = 0)
}
\arguments{
\item{fdataobj}{An object of class 'fdata' (functional response).}

\item{predictors}{A matrix of scalar predictors (n x p).}

\item{lambda}{Smoothing/penalty parameter (default 0).}
}
\value{
An object of class 'fosr' with components:
\item{intercept}{Intercept function mu(t) as fdata}
\item{beta}{Coefficient functions beta_j(t) as fdata}
\item{fitted}{Fitted functional values as fdata}
\item{residuals}{Residual functions as fdata}
\item{r.squared}{Global R-squared}
\item{r.squared.t}{Pointwise R-squared}
\item{gcv}{GCV criterion value}
\item{lambda}{Smoothing parameter used}
}
\description{
Functions for regression where the response is a function and predictors
are scalar variables.
Function-on-Scalar Regression (Penalized)
}
\details{
Fits a penalized function-on-scalar regression model where the response
is a function Y_i(t) and predictors are scalar: Y_i(t) = mu(t) + sum_j x_{ij} beta_j(t) + eps_i(t).
}
\seealso{
\code{\link{fosr.fpc}} for FPC-based FOSR,
\code{\link{fanova}} for functional ANOVA
}
