% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{fregre.pointwise}
\alias{fregre.pointwise}
\title{Pointwise Importance}
\usage{
fregre.pointwise(model)
}
\arguments{
\item{model}{A fitted \code{fregre.lm} or \code{fregre.logistic} model.}
}
\value{
A list with importance, importance_normalized,
component_importance, and score_variance.
}
\description{
Computes importance of each domain point for prediction.
}
