% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_fpca.R
\name{horiz.fpca}
\alias{horiz.fpca}
\title{Horizontal (Phase) FPCA}
\usage{
horiz.fpca(karcher, ncomp = 3)
}
\arguments{
\item{karcher}{A Karcher mean result from \code{\link{karcher.mean}}.}

\item{ncomp}{Number of principal components (default 3).}
}
\value{
An object of class 'horiz.fpca' with components:
\item{scores}{Matrix of FPC scores (n x ncomp)}
\item{eigenfunctions.psi}{Eigenfunctions in tangent space}
\item{eigenfunctions.gam}{Eigenfunctions as warp functions}
\item{eigenvalues}{Eigenvalues}
\item{cumulative.variance}{Cumulative proportion of variance}
\item{mean.psi}{Mean tangent vector}
\item{shooting.vectors}{Shooting vectors matrix}
\item{karcher}{The Karcher mean used}
}
\description{
Performs FPCA on the phase (warping) component after elastic alignment.
Captures timing variability independent of shape.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
km <- karcher.mean(fd)
h <- horiz.fpca(km, ncomp = 2)
h
}

}
