% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_fpca.R
\name{joint.fpca}
\alias{joint.fpca}
\title{Joint (Amplitude + Phase) FPCA}
\usage{
joint.fpca(karcher, ncomp = 3)
}
\arguments{
\item{karcher}{A Karcher mean result from \code{\link{karcher.mean}}.}

\item{ncomp}{Number of principal components (default 3).}
}
\value{
An object of class 'joint.fpca' with components:
\item{scores}{Matrix of joint FPC scores (n x ncomp)}
\item{eigenvalues}{Eigenvalues}
\item{cumulative.variance}{Cumulative proportion of variance}
\item{balance.c}{Balancing parameter between amplitude and phase}
\item{vert.component}{Vertical (amplitude) eigenvector components}
\item{horiz.component}{Horizontal (phase) eigenvector components}
\item{karcher}{The Karcher mean used}
}
\description{
Performs joint FPCA combining amplitude and phase components with a
balancing parameter.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
km <- karcher.mean(fd)
j <- joint.fpca(km, ncomp = 2)
j
}

}
