% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{rp.stat}
\alias{rp.stat}
\title{Random Projection Statistic}
\usage{
rp.stat(proj.x.ord, residuals, n.proj)
}
\arguments{
\item{proj.x.ord}{Integer vector of projection orderings.}

\item{residuals}{Numeric vector of residuals.}

\item{n.proj}{Number of random projections.}
}
\value{
A list with \code{cvm} and \code{ks} vectors of test statistics.
}
\description{
Computes Cramer-von Mises and Kolmogorov-Smirnov statistics for
random projection goodness-of-fit testing.
}
\examples{
\donttest{
set.seed(1)
n <- 10
n_proj <- 3
proj <- unlist(replicate(n_proj, sample(n), simplify = FALSE))
rp.stat(proj, rnorm(n), n.proj = n_proj)
}

}
