% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_basis.R
\name{smooth.basis.gcv}
\alias{smooth.basis.gcv}
\title{Penalized Basis Smoothing with GCV-Optimal Lambda}
\usage{
smooth.basis.gcv(
  fdataobj,
  type = c("bspline", "fourier"),
  nbasis = NULL,
  lfd.order = 2,
  log.lambda.range = c(-10, 5),
  n.grid = 50,
  period = NULL
)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{type}{Basis type: "bspline" or "fourier".}

\item{nbasis}{Number of basis functions. If NULL, defaults to min(ncol/4, 30).}

\item{lfd.order}{Order of the linear differential operator for the penalty.}

\item{log.lambda.range}{Range of log10(lambda) to search over.}

\item{n.grid}{Number of grid points for the lambda search.}

\item{period}{Period for Fourier basis.}
}
\value{
An object of class 'smooth.basis' (same as smooth.basis.fd).
}
\description{
Selects the smoothing parameter by minimizing the generalized
cross-validation (GCV) criterion over a grid of lambda values.
}
\examples{
\donttest{
fd <- fdata(matrix(rnorm(500), 50, 10), argvals = seq(0, 1, length.out = 10))
sm <- smooth.basis.gcv(fd, type = "bspline", nbasis = 5)
sm
}

}
