% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{smooth_basis_rust}
\alias{smooth_basis_rust}
\title{Penalized basis smoothing}
\usage{
smooth_basis_rust(data, argvals, basis_type, nbasis, lambda, lfd_order, period)
}
\description{
Penalized basis smoothing
}
