% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{spm.monitor}
\alias{spm.monitor}
\title{Monitor New Functional Data (Phase II)}
\usage{
spm.monitor(chart, newdata)
}
\arguments{
\item{chart}{An object of class \code{spm.chart} from \code{\link{spm.phase1}}.}

\item{newdata}{An object of class \code{fdata} with new observations to monitor.}
}
\value{
An object of class \code{spm.monitor} with components:
\describe{
\item{t2}{T-squared values for new observations}
\item{spe}{SPE values for new observations}
\item{t2.alarm}{Logical vector: TRUE where T-squared exceeds UCL}
\item{spe.alarm}{Logical vector: TRUE where SPE exceeds UCL}
\item{scores}{FPC score matrix for new observations}
\item{t2.ucl}{T-squared control limit}
\item{spe.ucl}{SPE control limit}
}
}
\description{
Projects new functional observations through a trained SPM chart and
checks whether T-squared and SPE statistics exceed control limits.
}
\examples{
\donttest{
# Build chart from in-control data
set.seed(1)
n <- 50; m <- 30
argvals <- seq(0, 1, length.out = m)
X <- matrix(rnorm(n * m), n, m)
fd <- fdata(X, argvals = argvals)
chart <- spm.phase1(fd, ncomp = 3)

# Monitor new data (with a shift)
X_new <- matrix(rnorm(10 * m) + 2, 10, m)
fd_new <- fdata(X_new, argvals = argvals)
mon <- spm.monitor(chart, fd_new)
mon
}

}
\seealso{
\code{\link{spm.phase1}} for building the chart
}
