% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{warp.smoothness}
\alias{warp.smoothness}
\title{Warping Smoothness}
\usage{
warp.smoothness(gamma, argvals)
}
\arguments{
\item{gamma}{Numeric vector of warping function values.}

\item{argvals}{Numeric vector of grid points.}
}
\value{
A scalar smoothness (bending energy) value.
}
\description{
Computes the bending energy of a warping function, measuring its roughness.
}
\examples{
t <- seq(0, 1, length.out = 20)
warp.smoothness(t^2, t)

}
